more raizing cleanups
