+++ src/mame/toaplan/raizing.cpp
@@ -3,23 +3,7 @@
 
 #include "emu.h"
 
-#include "toaplipt.h"
-#include "gp9001.h"
-
-#include "emupal.h"
-#include "screen.h"
-#include "speaker.h"
-#include "tilemap.h"
-
-#include "cpu/m68000/m68000.h"
-#include "cpu/nec/v25.h"
-#include "cpu/z80/z80.h"
-#include "machine/bankdev.h"
-#include "machine/eepromser.h"
-#include "machine/gen_latch.h"
-#include "sound/okim6295.h"
-#include "sound/ymopm.h"
-#include "sound/ymz280b.h"
+#include "raizing.h"
 
 // TODO: this is the old toaplan2.cpp header text, it could do with splitting between the new drivers
 
@@ -350,180 +334,6 @@ To Do / Unknowns:
 
 *****************************************************************************/
 
-
-class raizing_base_state : public driver_device
-{
-public:
-	raizing_base_state(const machine_config &mconfig, device_type type, const char *tag)
-		: driver_device(mconfig, type, tag)
-		, m_tx_videoram(*this, "tx_videoram")
-		, m_tx_lineselect(*this, "tx_lineselect")
-		, m_tx_linescroll(*this, "tx_linescroll")
-		, m_tx_gfxram(*this, "tx_gfxram")
-		, m_audiobank(*this, "audiobank")
-		, m_raizing_okibank{
-			{ *this, "raizing_okibank0_%u", 0U },
-			{ *this, "raizing_okibank1_%u", 0U } }
-		, m_shared_ram(*this, "shared_ram")
-		, m_maincpu(*this, "maincpu")
-		, m_audiocpu(*this, "audiocpu")
-		, m_vdp(*this, "gp9001")
-		, m_oki(*this, "oki%u", 1U)
-		, m_gfxdecode(*this, "gfxdecode")
-		, m_screen(*this, "screen")
-		, m_palette(*this, "palette")
-		, m_soundlatch(*this, "soundlatch%u", 1U)
-		, m_oki_rom(*this, "oki%u", 1U)
-	{ }
-
-protected:
-	u32 screen_update(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect);
-	virtual void device_post_load() override;
-
-	void bgaregga_common_video_start();
-
-	// used by everything
-	void create_tx_tilemap(int dx = 0, int dx_flipped = 0);
-
-	void raizing_z80_bankswitch_w(u8 data);
-	void raizing_oki_bankswitch_w(offs_t offset, u8 data);
-
-
-	// used by bgaregga and batrider
-	void install_raizing_okibank(int chip);
-
-
-	u32 screen_update_bootleg(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect);
-
-	DECLARE_MACHINE_RESET(bgaregga);
-	template<unsigned Chip> void raizing_oki(address_map &map) ATTR_COLD;
-
-	u8 shared_ram_r(offs_t offset) { return m_shared_ram[offset]; }
-	void shared_ram_w(offs_t offset, u8 data) { m_shared_ram[offset] = data; }
-
-	u32 screen_update_base(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect);
-	void screen_vblank(int state);
-
-	void tx_videoram_w(offs_t offset, u16 data, u16 mem_mask = ~0);
-	void tx_linescroll_w(offs_t offset, u16 data, u16 mem_mask = ~0);
-	TILE_GET_INFO_MEMBER(get_text_tile_info);
-
-	tilemap_t *m_tx_tilemap = nullptr;    /* Tilemap for extra-text-layer */
-	required_shared_ptr<u16> m_tx_videoram;
-	optional_shared_ptr<u16> m_tx_lineselect;
-	optional_shared_ptr<u16> m_tx_linescroll;
-	optional_shared_ptr<u16> m_tx_gfxram;
-	optional_memory_bank m_audiobank; // batrider and bgaregga
-	optional_memory_bank_array<8> m_raizing_okibank[2];
-	void coin_w(u8 data);
-	void reset(int state);
-
-	optional_shared_ptr<u8> m_shared_ram; // 8 bit RAM shared between 68K and sound CPU
-
-	required_device<m68000_base_device> m_maincpu;
-	optional_device<cpu_device> m_audiocpu;
-	required_device<gp9001vdp_device> m_vdp;
-	optional_device_array<okim6295_device, 2> m_oki;
-	optional_device<gfxdecode_device> m_gfxdecode;
-	required_device<screen_device> m_screen;
-	required_device<palette_device> m_palette;
-	optional_device_array<generic_latch_8_device, 4> m_soundlatch; // tekipaki, batrider, bgaregga, batsugun
-	optional_region_ptr_array<u8, 2> m_oki_rom;
-	bitmap_ind8 m_custom_priority_bitmap;
-};
-
-class batrider_state : public raizing_base_state
-{
-public:
-	batrider_state(const machine_config &mconfig, device_type type, const char *tag)
-		: raizing_base_state(mconfig, type, tag)
-		, m_dma_space(*this, "dma_space")
-		, m_mainram(*this, "mainram")
-		, m_z80_rom(*this, "audiocpu")
-	{ }
-
-	void batrider(machine_config &config);
-
-	void init_batrider();
-
-protected:
-	virtual void machine_start() override ATTR_COLD;
-	virtual void video_start() override ATTR_COLD;
-
-	void batrider_68k_mem(address_map &map) ATTR_COLD;
-	void batrider_dma_mem(address_map &map) ATTR_COLD;
-	void batrider_sound_z80_mem(address_map &map) ATTR_COLD;
-	void batrider_sound_z80_port(address_map &map) ATTR_COLD;
-
-	u16 batrider_z80_busack_r();
-	void batrider_z80_busreq_w(u8 data);
-	u16 batrider_z80rom_r(offs_t offset);
-	void batrider_soundlatch_w(u8 data);
-	void batrider_soundlatch2_w(u8 data);
-	void batrider_unknown_sound_w(u16 data);
-	void batrider_clear_sndirq_w(u16 data);
-	void batrider_sndirq_w(u8 data);
-	void batrider_clear_nmi_w(u8 data);
-	void batrider_tx_gfxram_w(offs_t offset, u16 data, u16 mem_mask = ~0);
-	void batrider_textdata_dma_w(u16 data);
-	void batrider_pal_text_dma_w(u16 data);
-	void batrider_objectbank_w(offs_t offset, u8 data);
-	void batrider_bank_cb(u8 layer, u32 &code);
-
-
-	required_device<address_map_bank_device> m_dma_space;
-	required_shared_ptr<u16> m_mainram;
-	optional_region_ptr<u8> m_z80_rom;
-
-	u16 m_gfxrom_bank[8]{};       /* Batrider object bank */
-	u8 m_sndirq_line = 0;        /* IRQ4 for batrider, IRQ2 for bbakraid */
-	u8 m_z80_busreq = 0;
-};
-
-
-class bbakraid_state : public batrider_state
-{
-public:
-	bbakraid_state(const machine_config &mconfig, device_type type, const char *tag)
-		: batrider_state(mconfig, type, tag)
-		, m_eepromout(*this, "EEPROMOUT")
-		, m_eeprom(*this, "eeprom")
-	{ }
-
-	void bbakraid(machine_config &config);
-
-	void init_bbakraid();
-
-protected:
-private:
-	void bbakraid_68k_mem(address_map &map) ATTR_COLD;
-	void bbakraid_sound_z80_mem(address_map &map) ATTR_COLD;
-	void bbakraid_sound_z80_port(address_map &map) ATTR_COLD;
-
-	INTERRUPT_GEN_MEMBER(bbakraid_snd_interrupt);
-
-	u16 bbakraid_eeprom_r();
-	void bbakraid_eeprom_w(u8 data);
-
-	optional_ioport m_eepromout;
-	optional_device<eeprom_serial_93cxx_device> m_eeprom;
-};
-
-
-class nprobowl_state : public batrider_state
-{
-public:
-	nprobowl_state(const machine_config &mconfig, device_type type, const char *tag)
-		: batrider_state(mconfig, type, tag)
-	{ }
-
-	void nprobowl(machine_config &config);
-
-protected:
-private:
-	void nprobowl_68k_mem(address_map &map) ATTR_COLD;
-};
-
 class bgaregga_state : public raizing_base_state
 {
 public:
@@ -536,6 +346,7 @@ public:
 	void init_bgaregga();
 
 protected:
+	virtual void machine_reset() override ATTR_COLD;
 	virtual void video_start() override ATTR_COLD;
 
 private:
@@ -557,6 +368,8 @@ public:
 protected:
 	virtual void video_start() override ATTR_COLD;
 
+private:
+	u32 screen_update_bootleg(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect);
 };
 
 
@@ -587,31 +400,6 @@ void raizing_base_state::reset(int state)
 	if (m_audiocpu != nullptr)
 		m_audiocpu->pulse_input_line(INPUT_LINE_RESET, attotime::zero);
 }
-
-void raizing_base_state::coin_w(u8 data) // MOVE TO DEVICE!
-{
-	/* +----------------+------ Bits 7-5 not used ------+--------------+ */
-	/* | Coin Lockout 2 | Coin Lockout 1 | Coin Count 2 | Coin Count 1 | */
-	/* |     Bit 3      |     Bit 2      |     Bit 1    |     Bit 0    | */
-
-	if (data & 0x0f)
-	{
-		machine().bookkeeping().coin_lockout_w(0, BIT(~data, 2));
-		machine().bookkeeping().coin_lockout_w(1, BIT(~data, 3));
-		machine().bookkeeping().coin_counter_w(0, BIT( data, 0));
-		machine().bookkeeping().coin_counter_w(1, BIT( data, 1));
-	}
-	else
-	{
-		machine().bookkeeping().coin_lockout_global_w(1);    // Lock all coin slots
-	}
-	if (data & 0xf0)
-	{
-		logerror("Writing unknown upper bits (%02x) to coin control\n",data);
-	}
-}
-
-
 
 TILE_GET_INFO_MEMBER(raizing_base_state::get_text_tile_info)
 {
@@ -688,56 +476,6 @@ u32 raizing_base_state::screen_update(screen_device &screen, bitmap_ind16 &bitma
 }
 
 
-void batrider_state::batrider_tx_gfxram_w(offs_t offset, u16 data, u16 mem_mask)
-{
-	/*** Dynamic GFX decoding for Batrider / Battle Bakraid ***/
-
-	const u16 oldword = m_tx_gfxram[offset];
-
-	if (oldword != data)
-	{
-		COMBINE_DATA(&m_tx_gfxram[offset]);
-		m_gfxdecode->gfx(0)->mark_dirty(offset/16);
-	}
-}
-
-void batrider_state::batrider_textdata_dma_w(u16 data)
-{
-	/*** Dynamic Text GFX decoding for Batrider ***/
-	/*** Only done once during start-up ***/
-	m_dma_space->set_bank(1);
-	for (int i = 0; i < (0x8000 >> 1); i++)
-	{
-		m_dma_space->write16(i, m_mainram[i]);
-	}
-}
-
-void batrider_state::batrider_pal_text_dma_w(u16 data)
-{
-	// FIXME: In batrider and bbakraid, the text layer and palette RAM
-	// are probably DMA'd from main RAM by writing here at every vblank,
-	// rather than being directly accessible to the 68K like the other games
-	m_dma_space->set_bank(0);
-	for (int i = 0; i < (0x3400 >> 1); i++)
-	{
-		m_dma_space->write16(i, m_mainram[i]);
-	}
-}
-
-void batrider_state::batrider_objectbank_w(offs_t offset, u8 data)
-{
-	data &= 0xf;
-	if (m_gfxrom_bank[offset] != data)
-	{
-		m_gfxrom_bank[offset] = data;
-		m_vdp->set_dirty();
-	}
-}
-
-void batrider_state::batrider_bank_cb(u8 layer, u32 &code)
-{
-	code = (m_gfxrom_bank[code >> 15] << 15) | (code & 0x7fff);
-}
 
 void raizing_base_state::create_tx_tilemap(int dx, int dx_flipped)
 {
@@ -777,26 +515,10 @@ void bgaregga_bootleg_state::video_start()
 	create_tx_tilemap(4, 4);
 }
 
-void batrider_state::video_start()
-{
-	m_screen->register_screen_bitmap(m_custom_priority_bitmap);
-	m_vdp->custom_priority_bitmap = &m_custom_priority_bitmap;
-
-	m_vdp->disable_sprite_buffer(); // disable buffering on this game
-
-	/* Create the Text tilemap for this game */
-	m_gfxdecode->gfx(0)->set_source(reinterpret_cast<u8 *>(m_tx_gfxram.target()));
-
-	create_tx_tilemap(0x1d4, 0x16b);
-
-	/* Has special banking */
-	save_item(NAME(m_gfxrom_bank));
-}
 
 
 
-
-u32 raizing_base_state::screen_update_bootleg(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect)
+u32 bgaregga_bootleg_state::screen_update_bootleg(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect)
 {
 	screen_update_base(screen, bitmap, cliprect);
 	m_tx_tilemap->draw(screen, bitmap, cliprect, 0);
@@ -1170,241 +892,16 @@ static INPUT_PORTS_START( bgareggabl )
 INPUT_PORTS_END
 
 
-static INPUT_PORTS_START( batrider )
-	PORT_START("IN")        // Player Inputs
-	PORT_BIT( 0x0001, IP_ACTIVE_HIGH, IPT_JOYSTICK_UP ) PORT_8WAY PORT_PLAYER(1)
-	PORT_BIT( 0x0002, IP_ACTIVE_HIGH, IPT_JOYSTICK_DOWN ) PORT_8WAY PORT_PLAYER(1)
-	PORT_BIT( 0x0004, IP_ACTIVE_HIGH, IPT_JOYSTICK_LEFT ) PORT_8WAY PORT_PLAYER(1)
-	PORT_BIT( 0x0008, IP_ACTIVE_HIGH, IPT_JOYSTICK_RIGHT ) PORT_8WAY PORT_PLAYER(1)
-	PORT_BIT( 0x0010, IP_ACTIVE_HIGH, IPT_BUTTON1 ) PORT_PLAYER(1)
-	PORT_BIT( 0x0020, IP_ACTIVE_HIGH, IPT_BUTTON2 ) PORT_PLAYER(1)
-	PORT_BIT( 0x0040, IP_ACTIVE_HIGH, IPT_BUTTON3 ) PORT_PLAYER(1)
-	PORT_BIT( 0x0100, IP_ACTIVE_HIGH, IPT_JOYSTICK_UP ) PORT_8WAY PORT_PLAYER(2)
-	PORT_BIT( 0x0200, IP_ACTIVE_HIGH, IPT_JOYSTICK_DOWN ) PORT_8WAY PORT_PLAYER(2)
-	PORT_BIT( 0x0400, IP_ACTIVE_HIGH, IPT_JOYSTICK_LEFT ) PORT_8WAY PORT_PLAYER(2)
-	PORT_BIT( 0x0800, IP_ACTIVE_HIGH, IPT_JOYSTICK_RIGHT ) PORT_8WAY PORT_PLAYER(2)
-	PORT_BIT( 0x1000, IP_ACTIVE_HIGH, IPT_BUTTON1 ) PORT_PLAYER(2)
-	PORT_BIT( 0x2000, IP_ACTIVE_HIGH, IPT_BUTTON2 ) PORT_PLAYER(2)
-	PORT_BIT( 0x4000, IP_ACTIVE_HIGH, IPT_BUTTON3 ) PORT_PLAYER(2)
-	PORT_BIT( 0x8080, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-
-	PORT_START("DSW")       // DSWA and DSWB
-	PORT_SERVICE_DIPLOC(0x0001, IP_ACTIVE_HIGH, "SW1:!1")
-	PORT_DIPNAME( 0x0002,   0x0000, "Credits to Start" )    PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)    PORT_DIPLOCATION("SW1:!2")
-	PORT_DIPSETTING(        0x0000, "1" )                   PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)
-	PORT_DIPSETTING(        0x0002, "2" )                   PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)
-	PORT_DIPNAME( 0x0002,   0x0000, "Joystick Mode" )       PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)       PORT_DIPLOCATION("SW1:!2")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Normal ) )     PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)
-	PORT_DIPSETTING(        0x0002, "90 degrees ACW" )      PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)
-	PORT_DIPNAME( 0x001c,   0x0000, DEF_STR( Coin_A ) )     PORT_DIPLOCATION("SW1:!3,!4,!5")
-	PORT_DIPSETTING(        0x0018, DEF_STR( 4C_1C ) )
-	PORT_DIPSETTING(        0x0014, DEF_STR( 3C_1C ) )
-	PORT_DIPSETTING(        0x0010, DEF_STR( 2C_1C ) )
-	PORT_DIPSETTING(        0x0000, DEF_STR( 1C_1C ) )
-	PORT_DIPSETTING(        0x0004, DEF_STR( 1C_2C ) )
-	PORT_DIPSETTING(        0x0008, DEF_STR( 1C_3C ) )
-	PORT_DIPSETTING(        0x000c, DEF_STR( 1C_4C ) )
-	PORT_DIPSETTING(        0x001c, DEF_STR( Free_Play ) )
-	PORT_DIPNAME( 0x00e0,   0x0000, DEF_STR( Coin_B ) )     PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)    PORT_DIPLOCATION("SW1:!6,!7,!8")
-	PORT_DIPSETTING(        0x00c0, DEF_STR( 4C_1C ) )      PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)
-	PORT_DIPSETTING(        0x00a0, DEF_STR( 3C_1C ) )      PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)
-	PORT_DIPSETTING(        0x0080, DEF_STR( 2C_1C ) )      PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)
-	PORT_DIPSETTING(        0x0000, DEF_STR( 1C_1C ) )      PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)
-//  PORT_DIPSETTING(        0x00e0, DEF_STR( 1C_1C ) )      PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)
-	PORT_DIPSETTING(        0x0020, DEF_STR( 1C_2C ) )      PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)
-	PORT_DIPSETTING(        0x0040, DEF_STR( 1C_3C ) )      PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)
-	PORT_DIPSETTING(        0x0060, DEF_STR( 1C_4C ) )      PORT_CONDITION("DSW", 0x001c, NOTEQUALS, 0x001c)
-	// When Coin_A is set to Free_Play, Coin_A becomes Coin_A and Coin_B, and the following dips occur
-	PORT_DIPNAME( 0x0020,   0x0000, "Hit Score" )           PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)   PORT_DIPLOCATION("SW1:!6")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Off ) )        PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)
-	PORT_DIPSETTING(        0x0020, DEF_STR( On ) )         PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)
-	PORT_DIPNAME( 0x0040,   0x0000, "Sound Effect" )        PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)   PORT_DIPLOCATION("SW1:!7")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Off ) )        PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)
-	PORT_DIPSETTING(        0x0040, DEF_STR( On ) )         PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)
-	PORT_DIPNAME( 0x0080,   0x0000, "Music" )               PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)   PORT_DIPLOCATION("SW1:!8")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Off ) )        PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)
-	PORT_DIPSETTING(        0x0080, DEF_STR( On ) )         PORT_CONDITION("DSW", 0x001c, EQUALS, 0x001c)
-	PORT_DIPNAME( 0x0300,   0x0000, DEF_STR( Difficulty ) ) PORT_DIPLOCATION("SW2:!1,!2")
-	PORT_DIPSETTING(        0x0100, DEF_STR( Easy ) )
-	PORT_DIPSETTING(        0x0000, DEF_STR( Normal ) )
-	PORT_DIPSETTING(        0x0200, DEF_STR( Hard ) )
-	PORT_DIPSETTING(        0x0300, DEF_STR( Very_Hard ) )
-	PORT_DIPNAME( 0x0c00,   0x0000, "Timer" )               PORT_DIPLOCATION("SW2:!3,!4")
-	PORT_DIPSETTING(        0x0c00, DEF_STR( Highest ) )
-	PORT_DIPSETTING(        0x0800, DEF_STR( High ) )
-	PORT_DIPSETTING(        0x0000, DEF_STR( Normal ) )
-	PORT_DIPSETTING(        0x0400, DEF_STR( Low ) )
-	PORT_DIPNAME( 0x3000,   0x0000, DEF_STR( Lives ) )      PORT_DIPLOCATION("SW2:!5,!6")
-	PORT_DIPSETTING(        0x3000, "1" )
-	PORT_DIPSETTING(        0x2000, "2" )
-	PORT_DIPSETTING(        0x0000, "3" )
-	PORT_DIPSETTING(        0x1000, "4" )
-	PORT_DIPNAME( 0xc000,   0x0000, DEF_STR( Bonus_Life ) ) PORT_DIPLOCATION("SW2:!7,!8")
-	PORT_DIPSETTING(        0xc000, DEF_STR( None ) )
-	PORT_DIPSETTING(        0x8000, "Every 2000k" )
-	PORT_DIPSETTING(        0x0000, "Every 1500k" )
-	PORT_DIPSETTING(        0x4000, "Every 1000k" )
-
-	PORT_START("SYS-DSW")   // Coin/System and DSWC
-	PORT_BIT( 0x0001, IP_ACTIVE_HIGH, IPT_SERVICE1 )
-	PORT_BIT( 0x0002, IP_ACTIVE_HIGH, IPT_UNKNOWN)
-	TOAPLAN_TEST_SWITCH( 0x0004, IP_ACTIVE_HIGH )
-	PORT_BIT( 0x0008, IP_ACTIVE_HIGH, IPT_COIN1 )
-	PORT_BIT( 0x0010, IP_ACTIVE_HIGH, IPT_COIN2 )
-	PORT_BIT( 0x0020, IP_ACTIVE_HIGH, IPT_START1 )
-	PORT_BIT( 0x0040, IP_ACTIVE_HIGH, IPT_START2 )
-	PORT_BIT( 0x0080, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_DIPNAME( 0x0100,   0x0000, DEF_STR( Flip_Screen ) )    PORT_DIPLOCATION("SW3:!1")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x0100, DEF_STR( On ) )
-	PORT_DIPNAME( 0x0200,   0x0000, DEF_STR( Demo_Sounds ) )    PORT_DIPLOCATION("SW3:!2")
-	PORT_DIPSETTING(        0x0200, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x0000, DEF_STR( On ) )
-	PORT_DIPNAME( 0x0400,   0x0000, "Stage Edit" )              PORT_DIPLOCATION("SW3:!3")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x0400, DEF_STR( On ) )
-	PORT_DIPNAME( 0x0800,   0x0000, DEF_STR( Allow_Continue ) ) PORT_DIPLOCATION("SW3:!4")
-	PORT_DIPSETTING(        0x0800, DEF_STR( No ) )
-	PORT_DIPSETTING(        0x0000, DEF_STR( Yes ) )
-	PORT_DIPNAME( 0x1000,   0x0000, "Invulnerability (Cheat)" )         PORT_DIPLOCATION("SW3:!5")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x1000, DEF_STR( On ) )
-	// These dips are shown only when Coin_A is set to Free_Play, but they work in normal play mode too
-	PORT_DIPNAME( 0x2000,   0x0000, "Guest Players" )           PORT_DIPLOCATION("SW3:!6")
-	PORT_DIPSETTING(        0x2000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x0000, DEF_STR( On ) )
-	PORT_DIPNAME( 0x4000,   0x0000, "Player Select" )           PORT_DIPLOCATION("SW3:!7")
-	PORT_DIPSETTING(        0x4000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x0000, DEF_STR( On ) )
-	PORT_DIPNAME( 0x8000,   0x0000, "Special Course" )          PORT_DIPLOCATION("SW3:!8")
-	PORT_DIPSETTING(        0x8000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x0000, DEF_STR( On ) )
-INPUT_PORTS_END
-
-
-static INPUT_PORTS_START( batriderj )
-	PORT_INCLUDE( batrider )
-
-	PORT_MODIFY("SYS-DSW")  // Coin/System and DSWC
-	// These dips are shown only when Coin_A is set to Free_Play, but they work in normal play mode too
-	PORT_DIPNAME( 0x2000,   0x0000, "Guest Players" )           PORT_DIPLOCATION("SW3:!6")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x2000, DEF_STR( On ) )
-	PORT_DIPNAME( 0x4000,   0x0000, "Player Select" )           PORT_DIPLOCATION("SW3:!7")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x4000, DEF_STR( On ) )
-	PORT_DIPNAME( 0x8000,   0x0000, "Special Course" )          PORT_DIPLOCATION("SW3:!8")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x8000, DEF_STR( On ) )
-INPUT_PORTS_END
-
-
-static INPUT_PORTS_START( bbakraid )
-	PORT_INCLUDE( batrider )
-
-	PORT_MODIFY("DSW")       // DSWA and DSWB
-	PORT_DIPNAME( 0xc000,   0x0000, DEF_STR( Bonus_Life ) ) PORT_DIPLOCATION("SW2:!7,!8")
-	PORT_DIPSETTING(        0xc000, DEF_STR( None ) )
-	PORT_DIPSETTING(        0x8000, "Every 4000k" )
-	PORT_DIPSETTING(        0x4000, "Every 3000k" )
-	PORT_DIPSETTING(        0x0000, "Every 2000k" )
-
-	PORT_MODIFY("SYS-DSW")   // Coin/System and DSW-3
-	PORT_DIPNAME( 0x2000,   0x0000, "Save Scores" )             PORT_DIPLOCATION("SW3:!6")
-	PORT_DIPSETTING(        0x2000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x0000, DEF_STR( On ) )
-	PORT_DIPNAME( 0x4000,   0x0000, DEF_STR( Unused ) )         PORT_DIPLOCATION("SW3:!7")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x4000, DEF_STR( On ) )
-	PORT_DIPNAME( 0x8000,   0x0000, DEF_STR( Unused ) )         PORT_DIPLOCATION("SW3:!8")
-	PORT_DIPSETTING(        0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(        0x8000, DEF_STR( On ) )
-
-	PORT_START( "EEPROMOUT" )
-	PORT_BIT( 0x0001, IP_ACTIVE_HIGH, IPT_OUTPUT ) PORT_WRITE_LINE_DEVICE_MEMBER("eeprom", FUNC(eeprom_serial_93cxx_device::cs_write))
-	PORT_BIT( 0x0004, IP_ACTIVE_HIGH, IPT_OUTPUT ) PORT_WRITE_LINE_DEVICE_MEMBER("eeprom", FUNC(eeprom_serial_93cxx_device::di_write))
-	PORT_BIT( 0x0008, IP_ACTIVE_HIGH, IPT_OUTPUT ) PORT_WRITE_LINE_DEVICE_MEMBER("eeprom", FUNC(eeprom_serial_93cxx_device::clk_write))
-INPUT_PORTS_END
-
-
-static INPUT_PORTS_START( nprobowl )
-	PORT_START("IN")        // Player Inputs
-	PORT_BIT( 0x0001, IP_ACTIVE_HIGH, IPT_COIN1 )
-	PORT_BIT( 0x0002, IP_ACTIVE_HIGH, IPT_SERVICE1 )
-	PORT_BIT( 0x0004, IP_ACTIVE_HIGH, IPT_START1 )
-	PORT_BIT( 0x0008, IP_ACTIVE_HIGH, IPT_START2 )
-	PORT_BIT( 0x0010, IP_ACTIVE_HIGH, IPT_BUTTON1 ) PORT_NAME("Set (Relay)")
-	PORT_BIT( 0x0020, IP_ACTIVE_HIGH, IPT_UNKNOWN ) // no effect in test mode
-	PORT_BIT( 0x0040, IP_ACTIVE_HIGH, IPT_BUTTON2 ) PORT_NAME("Gutter L")
-	PORT_BIT( 0x0080, IP_ACTIVE_HIGH, IPT_BUTTON3 ) PORT_NAME("Gutter R")
-	PORT_BIT( 0x0100, IP_ACTIVE_HIGH, IPT_UNKNOWN ) // no effect in test mode
-	PORT_BIT( 0x0200, IP_ACTIVE_HIGH, IPT_BUTTON4 ) PORT_NAME("Ballout")
-	PORT_BIT( 0x0400, IP_ACTIVE_HIGH, IPT_BUTTON5 ) PORT_NAME("Ballpass")
-	PORT_BIT( 0x0800, IP_ACTIVE_HIGH, IPT_BUTTON6 ) PORT_NAME("Mot Home")
-	PORT_BIT( 0x1000, IP_ACTIVE_HIGH, IPT_UNKNOWN ) // no effect in test mode
-	PORT_BIT( 0x2000, IP_ACTIVE_HIGH, IPT_UNKNOWN ) // no effect in test mode
-	PORT_BIT( 0x4000, IP_ACTIVE_HIGH, IPT_UNKNOWN ) // no effect in test mode
-	PORT_BIT( 0x8000, IP_ACTIVE_HIGH, IPT_UNKNOWN ) // no effect in test mode
-
-	PORT_START("DSW")       // SW0323 and SW0324
-	PORT_SERVICE_DIPLOC(   0x0001, IP_ACTIVE_HIGH, "SW0323:!1")
-	PORT_DIPUNKNOWN_DIPLOC(0x0002, 0x0000, "SW0323:!2")
-	PORT_DIPNAME(          0x0004, 0x0004, DEF_STR( Demo_Sounds ) ) PORT_DIPLOCATION("SW0323:!3")
-	PORT_DIPSETTING(       0x0004, DEF_STR( On ) )
-	PORT_DIPSETTING(       0x0000, DEF_STR( Off ) )
-	PORT_DIPUNKNOWN_DIPLOC(0x0008, 0x0000, "SW0323:!4")
-	PORT_DIPNAME(          0x0070, 0x0000, DEF_STR( Coinage ) )  PORT_DIPLOCATION("SW0323:!5,!6,!7")
-	PORT_DIPSETTING(       0x0070, DEF_STR( 5C_1C ) )
-	PORT_DIPSETTING(       0x0060, DEF_STR( 4C_1C ) )
-	PORT_DIPSETTING(       0x0050, DEF_STR( 3C_1C ) )
-	PORT_DIPSETTING(       0x0040, DEF_STR( 2C_1C ) )
-	PORT_DIPSETTING(       0x0000, DEF_STR( 1C_1C ) )
-	PORT_DIPSETTING(       0x0010, DEF_STR( 1C_2C ) )
-	PORT_DIPSETTING(       0x0020, DEF_STR( 1C_3C ) )
-	PORT_DIPSETTING(       0x0030, DEF_STR( 1C_4C ) )
-	PORT_DIPNAME(          0x0080, 0x0000, DEF_STR( Free_Play ) )  PORT_DIPLOCATION("SW0323:!8")
-	PORT_DIPSETTING(       0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(       0x0080, DEF_STR( On ) )
-	PORT_DIPUNKNOWN_DIPLOC(0x0100, 0x0000, "SW0324:!1")
-	PORT_DIPUNKNOWN_DIPLOC(0x0200, 0x0000, "SW0324:!2")
-	PORT_DIPUNKNOWN_DIPLOC(0x0400, 0x0000, "SW0324:!3")
-	PORT_DIPUNKNOWN_DIPLOC(0x0800, 0x0000, "SW0324:!4")
-	PORT_DIPUNKNOWN_DIPLOC(0x1000, 0x0000, "SW0324:!5")
-	PORT_DIPUNKNOWN_DIPLOC(0x2000, 0x0000, "SW0324:!6")
-	PORT_DIPUNKNOWN_DIPLOC(0x4000, 0x0000, "SW0324:!7")
-	PORT_DIPUNKNOWN_DIPLOC(0x8000, 0x0000, "SW0324:!8")
-
-	PORT_START("UNK")   // ??
-	PORT_BIT( 0x0001, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x0002, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x0004, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x0008, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x0010, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x0020, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x0040, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x0080, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x0100, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x0200, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x0400, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x0800, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x1000, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x2000, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x4000, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-	PORT_BIT( 0x8000, IP_ACTIVE_HIGH, IPT_UNKNOWN )
-INPUT_PORTS_END
-
 
 void raizing_base_state::raizing_z80_bankswitch_w(u8 data)
 {
 	m_audiobank->set_entry(data & 0x0f);
 }
 
-
 // bgaregga and batrider don't actually have a NMK112, but rather a GAL
 // programmed to bankswitch the sound ROMs in a similar fashion.
 // it may not be a coincidence that the composer and sound designer for
 // these two games, Manabu "Santaruru" Namiki, came to Raizing from NMK...
-
 void raizing_base_state::raizing_oki_bankswitch_w(offs_t offset, u8 data)
 {
 	m_raizing_okibank[(offset & 4) >> 2][offset & 3]->set_entry(data & 0xf);
@@ -1415,7 +912,6 @@ void raizing_base_state::raizing_oki_bankswitch_w(offs_t offset, u8 data)
 	m_raizing_okibank[(offset & 4) >> 2][4 + (offset & 3)]->set_entry(data & 0xf);
 }
 
-
 u8 bgaregga_state::bgaregga_E01D_r()
 {
 	// the Z80 reads this address during its IRQ routine,
@@ -1423,109 +919,14 @@ u8 bgaregga_state::bgaregga_E01D_r()
 	return m_soundlatch[0]->pending_r() ? 0 : 1;
 }
 
-
-u16 batrider_state::batrider_z80_busack_r()
-{
-	// Bit 0x01 returns the status of BUSAK from the Z80.
-	// These accesses are made when the 68K wants to read the Z80
-	// ROM code. Failure to return the correct status incurrs a Sound Error.
-
-	return m_z80_busreq;    // Loop BUSRQ to BUSAK
-}
-
-
-void batrider_state::batrider_z80_busreq_w(u8 data)
-{
-	m_z80_busreq = (data & 0x01);   // see batrider_z80_busack_r above
-}
-
-
-u16 batrider_state::batrider_z80rom_r(offs_t offset)
-{
-	return m_z80_rom[offset];
-}
-
-// these two latches are always written together, via a single move.l instruction
-void batrider_state::batrider_soundlatch_w(u8 data)
-{
-	m_soundlatch[0]->write(data & 0xff);
-	m_audiocpu->set_input_line(INPUT_LINE_NMI, ASSERT_LINE);
-}
-
-
-void batrider_state::batrider_soundlatch2_w(u8 data)
-{
-	m_soundlatch[1]->write(data & 0xff);
-	m_audiocpu->set_input_line(INPUT_LINE_NMI, ASSERT_LINE);
-}
-
-void batrider_state::batrider_unknown_sound_w(u16 data)
-{
-	// the 68K writes here when it wants a sound acknowledge IRQ from the Z80
-	// for bbakraid this is on every sound command; for batrider, only on certain commands
-}
-
-
-void batrider_state::batrider_clear_sndirq_w(u16 data)
-{
-	// not sure whether this is correct
-	// the 68K writes here during the sound IRQ handler, and nowhere else...
-	m_maincpu->set_input_line(m_sndirq_line, CLEAR_LINE);
-}
-
-
-void batrider_state::batrider_sndirq_w(u8 data)
-{
-	// if batrider_clear_sndirq_w() is correct, should this be ASSERT_LINE?
-	m_maincpu->set_input_line(m_sndirq_line, HOLD_LINE);
-}
-
-
-void batrider_state::batrider_clear_nmi_w(u8 data)
-{
-	m_audiocpu->set_input_line(INPUT_LINE_NMI, CLEAR_LINE);
-}
-
-
-u16 bbakraid_state::bbakraid_eeprom_r()
-{
-	// Bit 0x01 returns the status of BUSAK from the Z80.
-	// BUSRQ is activated via bit 0x10 on the EEPROM write port.
-	// These accesses are made when the 68K wants to read the Z80
-	// ROM code. Failure to return the correct status incurrs a Sound Error.
-
-	u8 data;
-	data  = ((m_eeprom->do_read() & 0x01) << 4);
-	data |= ((m_z80_busreq >> 4) & 0x01);   // Loop BUSRQ to BUSAK
-
-	return data;
-}
-
-
-void bbakraid_state::bbakraid_eeprom_w(u8 data)
-{
-	if (data & ~0x1f)
-		logerror("CPU #0 PC:%06X - Unknown EEPROM data being written %02X\n",m_maincpu->pc(),data);
-
-	m_eepromout->write(data, 0xff);
-
-	m_z80_busreq = data & 0x10; // see bbakraid_eeprom_r above
-}
-
 void sstriker_state::shippumd_coin_w(u8 data)
 {
-	coin_w(data & ~0x10);
+	m_coincounter->coin_w(data & ~0x10);
 	m_oki[0]->set_rom_bank(BIT(data, 4));
 }
 
-INTERRUPT_GEN_MEMBER(bbakraid_state::bbakraid_snd_interrupt)
+void raizing_base_state::common_bgaregga_reset()
 {
-	device.execute().set_input_line(0, HOLD_LINE);
-}
-
-MACHINE_RESET_MEMBER(raizing_base_state, bgaregga)
-{
-	driver_device::machine_reset();
 	for (int chip = 0; chip < 2; chip++)
 	{
 		for (int i = 0; i < 8; i++)
@@ -1536,18 +937,17 @@ MACHINE_RESET_MEMBER(raizing_base_state, bgaregga)
 	}
 }
 
-void batrider_state::machine_start()
+void bgaregga_state::machine_reset()
 {
-	save_item(NAME(m_z80_busreq));
+	common_bgaregga_reset();
 }
-
 
 void sstriker_state::mahoudai_68k_mem(address_map &map)
 {
 	map(0x000000, 0x07ffff).rom();
 	map(0x100000, 0x10ffff).ram();
 	map(0x218000, 0x21bfff).rw(FUNC(sstriker_state::shared_ram_r), FUNC(sstriker_state::shared_ram_w)).umask16(0x00ff);
-	map(0x21c01d, 0x21c01d).w(FUNC(sstriker_state::coin_w));
+	map(0x21c01d, 0x21c01d).w("coincounter", FUNC(toaplan_coincounter_device::coin_w));
 	map(0x21c020, 0x21c021).portr("IN1");
 	map(0x21c024, 0x21c025).portr("IN2");
 	map(0x21c028, 0x21c029).portr("SYS");
@@ -1563,7 +963,6 @@ void sstriker_state::mahoudai_68k_mem(address_map &map)
 	map(0x503000, 0x5031ff).ram().w(FUNC(sstriker_state::tx_linescroll_w)).share(m_tx_linescroll);
 	map(0x503200, 0x503fff).ram();
 }
-
 
 void sstriker_state::shippumd_68k_mem(address_map &map)
 {
@@ -1588,13 +987,12 @@ void sstriker_state::shippumd_68k_mem(address_map &map)
 	map(0x503200, 0x503fff).ram();
 }
 
-
 void bgaregga_state::bgaregga_68k_mem(address_map &map)
 {
 	map(0x000000, 0x0fffff).rom();
 	map(0x100000, 0x10ffff).ram();
 	map(0x218000, 0x21bfff).rw(FUNC(bgaregga_state::shared_ram_r), FUNC(bgaregga_state::shared_ram_w)).umask16(0x00ff);
-	map(0x21c01d, 0x21c01d).w(FUNC(bgaregga_state::coin_w));
+	map(0x21c01d, 0x21c01d).w("coincounter", FUNC(toaplan_coincounter_device::coin_w));
 	map(0x21c020, 0x21c021).portr("IN1");
 	map(0x21c024, 0x21c025).portr("IN2");
 	map(0x21c028, 0x21c029).portr("SYS");
@@ -1611,107 +1009,14 @@ void bgaregga_state::bgaregga_68k_mem(address_map &map)
 	map(0x600001, 0x600001).w(m_soundlatch[0], FUNC(generic_latch_8_device::write));
 }
 
-
-void batrider_state::batrider_dma_mem(address_map &map)
-{
-	map(0x0000, 0x1fff).ram().w(FUNC(batrider_state::tx_videoram_w)).share(m_tx_videoram);
-	map(0x2000, 0x2fff).ram().w(m_palette, FUNC(palette_device::write16)).share("palette");
-	map(0x3000, 0x31ff).ram().share(m_tx_lineselect);
-	map(0x3200, 0x33ff).ram().w(FUNC(batrider_state::tx_linescroll_w)).share(m_tx_linescroll);
-	map(0x3400, 0x7fff).ram();
-	map(0x8000, 0xffff).ram().w(FUNC(batrider_state::batrider_tx_gfxram_w)).share(m_tx_gfxram);
-}
-
-
-void batrider_state::batrider_68k_mem(address_map &map)
-{
-	map(0x000000, 0x1fffff).rom();
-	// actually 200000 - 20ffff is probably all main RAM, and the text and palette RAM are written via DMA
-	map(0x200000, 0x207fff).ram().share(m_mainram);
-	map(0x208000, 0x20ffff).ram();
-	map(0x300000, 0x37ffff).r(FUNC(batrider_state::batrider_z80rom_r));
-	map(0x400000, 0x40000d).lrw16(
-							NAME([this](offs_t offset, u16 mem_mask) { return m_vdp->read(offset ^ (0xc/2), mem_mask); }),
-							NAME([this](offs_t offset, u16 data, u16 mem_mask) { m_vdp->write(offset ^ (0xc/2), data, mem_mask); }));
-	map(0x500000, 0x500001).portr("IN");
-	map(0x500002, 0x500003).portr("SYS-DSW");
-	map(0x500004, 0x500005).portr("DSW");
-	map(0x500006, 0x500007).r(m_vdp, FUNC(gp9001vdp_device::vdpcount_r));
-	map(0x500009, 0x500009).r(m_soundlatch[2], FUNC(generic_latch_8_device::read));
-	map(0x50000b, 0x50000b).r(m_soundlatch[3], FUNC(generic_latch_8_device::read));
-	map(0x50000c, 0x50000d).r(FUNC(batrider_state::batrider_z80_busack_r));
-	map(0x500011, 0x500011).w(FUNC(batrider_state::coin_w));
-	map(0x500021, 0x500021).w(FUNC(batrider_state::batrider_soundlatch_w));
-	map(0x500023, 0x500023).w(FUNC(batrider_state::batrider_soundlatch2_w));
-	map(0x500024, 0x500025).w(FUNC(batrider_state::batrider_unknown_sound_w));
-	map(0x500026, 0x500027).w(FUNC(batrider_state::batrider_clear_sndirq_w));
-	map(0x500061, 0x500061).w(FUNC(batrider_state::batrider_z80_busreq_w));
-	map(0x500080, 0x500081).w(FUNC(batrider_state::batrider_textdata_dma_w));
-	map(0x500082, 0x500083).w(FUNC(batrider_state::batrider_pal_text_dma_w));
-	map(0x5000c0, 0x5000cf).w(FUNC(batrider_state::batrider_objectbank_w)).umask16(0x00ff);
-}
-
-
-void bbakraid_state::bbakraid_68k_mem(address_map &map)
-{
-	map(0x000000, 0x1fffff).rom();
-	// actually 200000 - 20ffff is probably all main RAM, and the text and palette RAM are written via DMA
-	map(0x200000, 0x207fff).ram().share(m_mainram);
-	map(0x208000, 0x20ffff).ram();
-	map(0x300000, 0x33ffff).r(FUNC(bbakraid_state::batrider_z80rom_r));
-	map(0x400000, 0x40000d).lrw16(
-							NAME([this](offs_t offset, u16 mem_mask) { return m_vdp->read(offset ^ (0xc/2), mem_mask); }),
-							NAME([this](offs_t offset, u16 data, u16 mem_mask) { m_vdp->write(offset ^ (0xc/2), data, mem_mask); }));
-	map(0x500000, 0x500001).portr("IN");
-	map(0x500002, 0x500003).portr("SYS-DSW");
-	map(0x500004, 0x500005).portr("DSW");
-	map(0x500006, 0x500007).r(m_vdp, FUNC(gp9001vdp_device::vdpcount_r));
-	map(0x500009, 0x500009).w(FUNC(bbakraid_state::coin_w));
-	map(0x500011, 0x500011).r(m_soundlatch[2], FUNC(generic_latch_8_device::read));
-	map(0x500013, 0x500013).r(m_soundlatch[3], FUNC(generic_latch_8_device::read));
-	map(0x500015, 0x500015).w(FUNC(bbakraid_state::batrider_soundlatch_w));
-	map(0x500017, 0x500017).w(FUNC(bbakraid_state::batrider_soundlatch2_w));
-	map(0x500018, 0x500019).r(FUNC(bbakraid_state::bbakraid_eeprom_r));
-	map(0x50001a, 0x50001b).w(FUNC(bbakraid_state::batrider_unknown_sound_w));
-	map(0x50001c, 0x50001d).w(FUNC(bbakraid_state::batrider_clear_sndirq_w));
-	map(0x50001f, 0x50001f).w(FUNC(bbakraid_state::bbakraid_eeprom_w));
-	map(0x500080, 0x500081).w(FUNC(bbakraid_state::batrider_textdata_dma_w));
-	map(0x500082, 0x500083).w(FUNC(bbakraid_state::batrider_pal_text_dma_w));
-	map(0x5000c0, 0x5000cf).w(FUNC(bbakraid_state::batrider_objectbank_w)).umask16(0x00ff);
-}
-
-
-void nprobowl_state::nprobowl_68k_mem(address_map &map) // TODO: verify everything, implement oki banking
-{
-	map(0x000000, 0x0fffff).rom();
-	map(0x200000, 0x207fff).ram().share(m_mainram);
-	map(0x208000, 0x20ffff).ram();
-	map(0x400000, 0x40000d).lrw16(
-							NAME([this](offs_t offset, u16 mem_mask) { return m_vdp->read(offset ^ (0xc/2), mem_mask); }),
-							NAME([this](offs_t offset, u16 data, u16 mem_mask) { m_vdp->write(offset ^ (0xc/2), data, mem_mask); }));
-	map(0x500000, 0x500001).portr("IN");
-	map(0x500002, 0x500003).portr("UNK");
-	map(0x500004, 0x500005).portr("DSW");
-	//map(0x500010, 0x500011).w();
-	//map(0x500012, 0x500013).w();
-	map(0x500021, 0x500021).w(m_oki[0], FUNC(okim6295_device::write));
-	//map(0x500040, 0x500041).w();
-	//map(0x500042, 0x500043).w();
-	map(0x500060, 0x500061).lr16(NAME([this] () -> u16 { return machine().rand(); })); // TODO: Hack, probably checks something in the mechanical part, verify
-	map(0x500080, 0x500081).w(FUNC(nprobowl_state::batrider_textdata_dma_w));
-	map(0x500082, 0x500083).w(FUNC(nprobowl_state::batrider_pal_text_dma_w));
-}
-
-
 void sstriker_state::raizing_sound_z80_mem(address_map &map)
 {
 	map(0x0000, 0xbfff).rom();
 	map(0xc000, 0xdfff).ram().share(m_shared_ram);
 	map(0xe000, 0xe001).rw("ymsnd", FUNC(ym2151_device::read), FUNC(ym2151_device::write));
 	map(0xe004, 0xe004).rw(m_oki[0], FUNC(okim6295_device::read), FUNC(okim6295_device::write));
-	map(0xe00e, 0xe00e).w(FUNC(sstriker_state::coin_w));
+	map(0xe00e, 0xe00e).w("coincounter", FUNC(toaplan_coincounter_device::coin_w));
 }
-
 
 void bgaregga_state::bgaregga_sound_z80_mem(address_map &map)
 {
@@ -1727,87 +1032,9 @@ void bgaregga_state::bgaregga_sound_z80_mem(address_map &map)
 	map(0xe01d, 0xe01d).r(FUNC(bgaregga_state::bgaregga_E01D_r));
 }
 
-
-void batrider_state::batrider_sound_z80_mem(address_map &map)
-{
-	map(0x0000, 0x7fff).rom();
-	map(0x8000, 0xbfff).bankr(m_audiobank);
-	map(0xc000, 0xdfff).ram();
-}
-
-
-void batrider_state::batrider_sound_z80_port(address_map &map)
-{
-	map.global_mask(0xff);
-	map(0x40, 0x40).w(m_soundlatch[2], FUNC(generic_latch_8_device::write));
-	map(0x42, 0x42).w(m_soundlatch[3], FUNC(generic_latch_8_device::write));
-	map(0x44, 0x44).w(FUNC(batrider_state::batrider_sndirq_w));
-	map(0x46, 0x46).w(FUNC(batrider_state::batrider_clear_nmi_w));
-	map(0x48, 0x48).r(m_soundlatch[0], FUNC(generic_latch_8_device::read));
-	map(0x4a, 0x4a).r(m_soundlatch[1], FUNC(generic_latch_8_device::read));
-	map(0x80, 0x81).rw("ymsnd", FUNC(ym2151_device::read), FUNC(ym2151_device::write));
-	map(0x82, 0x82).rw(m_oki[0], FUNC(okim6295_device::read), FUNC(okim6295_device::write));
-	map(0x84, 0x84).rw(m_oki[1], FUNC(okim6295_device::read), FUNC(okim6295_device::write));
-	map(0x88, 0x88).w(FUNC(batrider_state::raizing_z80_bankswitch_w));
-	map(0xc0, 0xc6).w(FUNC(batrider_state::raizing_oki_bankswitch_w));
-}
-
-
-void bbakraid_state::bbakraid_sound_z80_mem(address_map &map)
-{
-	map(0x0000, 0xbfff).rom();     // No banking? ROM only contains code and data up to 0x28DC
-	map(0xc000, 0xffff).ram();
-}
-
-
-void bbakraid_state::bbakraid_sound_z80_port(address_map &map)
-{
-	map.global_mask(0xff);
-	map(0x40, 0x40).w(m_soundlatch[2], FUNC(generic_latch_8_device::write));
-	map(0x42, 0x42).w(m_soundlatch[3], FUNC(generic_latch_8_device::write));
-	map(0x44, 0x44).w(FUNC(bbakraid_state::batrider_sndirq_w));
-	map(0x46, 0x46).w(FUNC(bbakraid_state::batrider_clear_nmi_w));
-	map(0x48, 0x48).r(m_soundlatch[0], FUNC(generic_latch_8_device::read));
-	map(0x4a, 0x4a).r(m_soundlatch[1], FUNC(generic_latch_8_device::read));
-	map(0x80, 0x81).rw("ymz", FUNC(ymz280b_device::read), FUNC(ymz280b_device::write));
-}
-
-// similar as NMK112, but GAL-driven; NOT actual NMK112 is present
-template<unsigned Chip>
-void raizing_base_state::raizing_oki(address_map &map)
-{
-	map(0x00000, 0x000ff).bankr(m_raizing_okibank[Chip][0]);
-	map(0x00100, 0x001ff).bankr(m_raizing_okibank[Chip][1]);
-	map(0x00200, 0x002ff).bankr(m_raizing_okibank[Chip][2]);
-	map(0x00300, 0x003ff).bankr(m_raizing_okibank[Chip][3]);
-	map(0x00400, 0x0ffff).bankr(m_raizing_okibank[Chip][4]);
-	map(0x10000, 0x1ffff).bankr(m_raizing_okibank[Chip][5]);
-	map(0x20000, 0x2ffff).bankr(m_raizing_okibank[Chip][6]);
-	map(0x30000, 0x3ffff).bankr(m_raizing_okibank[Chip][7]);
-}
-
-#define XOR(a) WORD_XOR_LE(a)
-#define LOC(x) (x+XOR(0))
-
-static const gfx_layout batrider_tx_tilelayout =
-{
-	8,8,    /* 8x8 characters */
-	1024,   /* 1024 characters */
-	4,      /* 4 bits per pixel */
-	{ STEP4(0,1) },
-	{ XOR(0)*4, XOR(1)*4, XOR(2)*4, XOR(3)*4, XOR(4)*4, XOR(5)*4, XOR(6)*4, XOR(7)*4 },
-	{ STEP8(0,4*8) },
-	8*8*4
-};
-
 static GFXDECODE_START( gfx_textrom )
 	GFXDECODE_ENTRY( "text", 0, gfx_8x8x4_packed_msb, 64*16, 64 )
 GFXDECODE_END
-
-static GFXDECODE_START( gfx_batrider )
-	GFXDECODE_ENTRY( nullptr, 0, batrider_tx_tilelayout, 64*16, 64 )
-GFXDECODE_END
-
 
 void sstriker_state::mahoudai(machine_config &config)
 {
@@ -1818,6 +1045,8 @@ void sstriker_state::mahoudai(machine_config &config)
 
 	Z80(config, m_audiocpu, 32_MHz_XTAL/8);     // 4MHz, 32MHz Oscillator
 	m_audiocpu->set_addrmap(AS_PROGRAM, &sstriker_state::raizing_sound_z80_mem);
+
+	TOAPLAN_COINCOUNTER(config, m_coincounter, 0);
 
 	config.set_maximum_quantum(attotime::from_hz(600));
 
@@ -1862,9 +1091,9 @@ void bgaregga_state::bgaregga(machine_config &config)
 	Z80(config, m_audiocpu, 32_MHz_XTAL/8);     // 4MHz, 32MHz Oscillator
 	m_audiocpu->set_addrmap(AS_PROGRAM, &bgaregga_state::bgaregga_sound_z80_mem);
 
-	config.set_maximum_quantum(attotime::from_hz(6000));
+	TOAPLAN_COINCOUNTER(config, m_coincounter, 0);
 
-	MCFG_MACHINE_RESET_OVERRIDE(bgaregga_state,bgaregga)
+	config.set_maximum_quantum(attotime::from_hz(6000));
 
 	/* video hardware */
 	SCREEN(config, m_screen, SCREEN_TYPE_RASTER);
@@ -1903,154 +1132,6 @@ void bgaregga_bootleg_state::bgareggabl(machine_config &config)
 	m_screen->set_screen_update(FUNC(bgaregga_bootleg_state::screen_update_bootleg));
 }
 
-void batrider_state::batrider(machine_config &config)
-{
-	/* basic machine hardware */
-	M68000(config, m_maincpu, 32_MHz_XTAL/2);   // 16MHz, 32MHz Oscillator (verified)
-	m_maincpu->set_addrmap(AS_PROGRAM, &batrider_state::batrider_68k_mem);
-	m_maincpu->reset_cb().set(FUNC(batrider_state::reset));
-
-	Z80(config, m_audiocpu, 32_MHz_XTAL/6);     // 5.333MHz, 32MHz Oscillator (verified)
-	m_audiocpu->set_addrmap(AS_PROGRAM, &batrider_state::batrider_sound_z80_mem);
-	m_audiocpu->set_addrmap(AS_IO, &batrider_state::batrider_sound_z80_port);
-
-	config.set_maximum_quantum(attotime::from_hz(600));
-
-	MCFG_MACHINE_RESET_OVERRIDE(batrider_state,bgaregga)
-
-	ADDRESS_MAP_BANK(config, m_dma_space, 0);
-	m_dma_space->set_addrmap(0, &batrider_state::batrider_dma_mem);
-	m_dma_space->set_endianness(ENDIANNESS_BIG);
-	m_dma_space->set_data_width(16);
-	m_dma_space->set_addr_width(16);
-	m_dma_space->set_stride(0x8000);
-
-	/* video hardware */
-	SCREEN(config, m_screen, SCREEN_TYPE_RASTER);
-	m_screen->set_video_attributes(VIDEO_UPDATE_BEFORE_VBLANK);
-	m_screen->set_raw(27_MHz_XTAL/4, 432, 0, 320, 262, 0, 240);
-	m_screen->set_screen_update(FUNC(batrider_state::screen_update));
-	m_screen->screen_vblank().set(FUNC(batrider_state::screen_vblank));
-	m_screen->set_palette(m_palette);
-
-	GFXDECODE(config, m_gfxdecode, m_palette, gfx_batrider);
-	PALETTE(config, m_palette).set_format(palette_device::xBGR_555, gp9001vdp_device::VDP_PALETTE_LENGTH);
-
-	GP9001_VDP(config, m_vdp, 27_MHz_XTAL);
-	m_vdp->set_palette(m_palette);
-	m_vdp->set_tile_callback(FUNC(batrider_state::batrider_bank_cb));
-	m_vdp->vint_out_cb().set_inputline(m_maincpu, M68K_IRQ_2);
-
-	/* sound hardware */
-	SPEAKER(config, "mono").front_center();
-
-	// these two latches are always written together, via a single move.l instruction
-	GENERIC_LATCH_8(config, m_soundlatch[0]);
-	GENERIC_LATCH_8(config, m_soundlatch[1]);
-	GENERIC_LATCH_8(config, m_soundlatch[2]);
-	GENERIC_LATCH_8(config, m_soundlatch[3]);
-
-	YM2151(config, "ymsnd", 32_MHz_XTAL/8).add_route(ALL_OUTPUTS, "mono", 0.25); // 4MHz, 32MHz Oscillator (verified)
-
-	OKIM6295(config, m_oki[0], 32_MHz_XTAL/10, okim6295_device::PIN7_HIGH);
-	m_oki[0]->set_addrmap(0, &batrider_state::raizing_oki<0>);
-	m_oki[0]->add_route(ALL_OUTPUTS, "mono", 0.5);
-
-	OKIM6295(config, m_oki[1], 32_MHz_XTAL/10, okim6295_device::PIN7_LOW);
-	m_oki[1]->set_addrmap(0, &batrider_state::raizing_oki<1>);
-	m_oki[1]->add_route(ALL_OUTPUTS, "mono", 0.5);
-}
-
-
-void bbakraid_state::bbakraid(machine_config &config)
-{
-	/* basic machine hardware */
-	M68000(config, m_maincpu, 32_MHz_XTAL/2);   // 16MHz, 32MHz Oscillator
-	m_maincpu->set_addrmap(AS_PROGRAM, &bbakraid_state::bbakraid_68k_mem);
-	m_maincpu->reset_cb().set(FUNC(bbakraid_state::reset));
-
-	Z80(config, m_audiocpu, XTAL(32'000'000)/6);     /* 5.3333MHz , 32MHz Oscillator */
-	m_audiocpu->set_addrmap(AS_PROGRAM, &bbakraid_state::bbakraid_sound_z80_mem);
-	m_audiocpu->set_addrmap(AS_IO, &bbakraid_state::bbakraid_sound_z80_port);
-	m_audiocpu->set_periodic_int(FUNC(bbakraid_state::bbakraid_snd_interrupt), attotime::from_hz(XTAL(32'000'000) / 6 / 12000)); // sound CPU clock (divider unverified)
-
-	config.set_maximum_quantum(attotime::from_hz(600));
-
-	EEPROM_93C66_8BIT(config, m_eeprom);
-
-	ADDRESS_MAP_BANK(config, m_dma_space, 0);
-	m_dma_space->set_addrmap(0, &bbakraid_state::batrider_dma_mem);
-	m_dma_space->set_endianness(ENDIANNESS_BIG);
-	m_dma_space->set_data_width(16);
-	m_dma_space->set_addr_width(16);
-	m_dma_space->set_stride(0x8000);
-
-	/* video hardware */
-	SCREEN(config, m_screen, SCREEN_TYPE_RASTER);
-	m_screen->set_video_attributes(VIDEO_UPDATE_BEFORE_VBLANK);
-	m_screen->set_raw(27_MHz_XTAL/4, 432, 0, 320, 262, 0, 240);
-	m_screen->set_screen_update(FUNC(bbakraid_state::screen_update));
-	m_screen->screen_vblank().set(FUNC(bbakraid_state::screen_vblank));
-	m_screen->set_palette(m_palette);
-
-	GFXDECODE(config, m_gfxdecode, m_palette, gfx_batrider);
-	PALETTE(config, m_palette).set_format(palette_device::xBGR_555, gp9001vdp_device::VDP_PALETTE_LENGTH);
-
-	GP9001_VDP(config, m_vdp, 27_MHz_XTAL);
-	m_vdp->set_palette(m_palette);
-	m_vdp->set_tile_callback(FUNC(bbakraid_state::batrider_bank_cb));
-	m_vdp->vint_out_cb().set_inputline(m_maincpu, M68K_IRQ_1);
-
-	/* sound hardware */
-	SPEAKER(config, "mono").front_center();
-
-	// these two latches are always written together, via a single move.l instruction
-	GENERIC_LATCH_8(config, m_soundlatch[0]);
-	GENERIC_LATCH_8(config, m_soundlatch[1]);
-	GENERIC_LATCH_8(config, m_soundlatch[2]);
-	GENERIC_LATCH_8(config, m_soundlatch[3]);
-
-	YMZ280B(config, "ymz", 16.9344_MHz_XTAL).add_route(ALL_OUTPUTS, "mono", 1.0);
-	// IRQ not used ???  Connected to a test pin (TP082)
-}
-
-
-void nprobowl_state::nprobowl(machine_config &config)
-{
-	// basic machine hardware
-	M68000(config, m_maincpu, 32_MHz_XTAL / 2);   // 32MHz Oscillator, divisor not verified
-	m_maincpu->set_addrmap(AS_PROGRAM, &nprobowl_state::nprobowl_68k_mem);
-	m_maincpu->reset_cb().set(FUNC(nprobowl_state::reset));
-
-	ADDRESS_MAP_BANK(config, m_dma_space, 0);
-	m_dma_space->set_addrmap(0, &nprobowl_state::batrider_dma_mem);
-	m_dma_space->set_endianness(ENDIANNESS_BIG);
-	m_dma_space->set_data_width(16);
-	m_dma_space->set_addr_width(16);
-	m_dma_space->set_stride(0x8000);
-
-	// video hardware
-	SCREEN(config, m_screen, SCREEN_TYPE_RASTER);
-	m_screen->set_video_attributes(VIDEO_UPDATE_BEFORE_VBLANK);
-	m_screen->set_raw(27_MHz_XTAL/4, 432, 0, 320, 262, 0, 240);
-	m_screen->set_screen_update(FUNC(nprobowl_state::screen_update));
-	m_screen->screen_vblank().set(FUNC(nprobowl_state::screen_vblank));
-	m_screen->set_palette(m_palette);
-
-	GFXDECODE(config, m_gfxdecode, m_palette, gfx_batrider);
-	PALETTE(config, m_palette).set_format(palette_device::xBGR_555, gp9001vdp_device::VDP_PALETTE_LENGTH);
-
-	GP9001_VDP(config, m_vdp, 27_MHz_XTAL);
-	m_vdp->set_palette(m_palette);
-	m_vdp->vint_out_cb().set_inputline(m_maincpu, M68K_IRQ_2);
-
-	// sound hardware
-	SPEAKER(config, "mono").front_center();
-
-	OKIM6295(config, m_oki[0], 32_MHz_XTAL/8, okim6295_device::PIN7_HIGH).add_route(ALL_OUTPUTS, "mono", 1.0); // divisor not verified
-	// TODO: banking
-}
-
 
 void raizing_base_state::install_raizing_okibank(int chip)
 {
@@ -2077,21 +1158,6 @@ void bgaregga_state::init_bgaregga()
 }
 
 
-void batrider_state::init_batrider()
-{
-	u8 *Z80 = memregion("audiocpu")->base();
-
-	m_audiobank->configure_entries(0, 16, Z80, 0x4000);
-	install_raizing_okibank(0);
-	install_raizing_okibank(1);
-	m_sndirq_line = 4;
-}
-
-
-void bbakraid_state::init_bbakraid()
-{
-	m_sndirq_line = 2;
-}
 
 /* -------------------------- Raizing games ------------------------- */
 
@@ -2493,455 +1559,6 @@ ROM_START( bgareggablj ) // fixed on Japanese region
 	ROM_LOAD( "rom5.bin", 0x000000, 0x100000, CRC(f6d49863) SHA1(3a3c354852adad06e8a051511abfab7606bce382) )
 ROM_END
 
-/*
-   The region of Batrider is controlled by the first byte of rom prg0.u22
-   only sets which have been dumped from original PCBs are supported
-
-   original ROM labels have no indication of the region.
-
-   valid values are:
-    ( * denotes that this set has been found on an original PCB )
-
-   00 : Nippon *
-   01 : USA *
-   02 : Europe *
-   03 : Asia
-   04 : German (sic)
-   05 : Austria
-   06 : Belgium
-   07 : Denmark
-   08 : Finland
-   09 : France
-   0A : Great Britain
-   0B : Greece
-   0C : The Netherlands
-   0D : Italy
-   0E : Norway
-   0F : Portugal
-   10 : Spain
-   11 : Sweden
-   12 : Switzerland
-   13 : Australia
-   14 : New Zealand
-   15 : Taiwan
-   16 : Hong Kong
-   17 : Korea *
-   18 : China *
-   19 : No Region?
-   1A+: Invalid
-
-   For future reference, that would mean the following
-
-   ROM_LOAD16_BYTE( "prg0_nippon.u22",       0x000000, 0x080000, CRC(4f3fc729) SHA1(b32d51c254741b82171a86c271679522a7aefd34) )
-   ROM_LOAD16_BYTE( "prg0_usa.u22",          0x000000, 0x080000, CRC(2049d007) SHA1(f2a43547a6fc5083b03c1d59a85abbf6e1ce4cd9) )
-   ROM_LOAD16_BYTE( "prg0_europe.u22",       0x000000, 0x080000, CRC(91d3e975) SHA1(682885fc17f2424d475c282f239f42faf1aae076) )
-   ROM_LOAD16_BYTE( "prg0_asia.u22",         0x000000, 0x080000, CRC(fea5fe5b) SHA1(0008336ecd3886485ab1d9678880b1a0bc788f40) )
-   ROM_LOAD16_BYTE( "prg0_german.u22",       0x000000, 0x080000, CRC(29969dd0) SHA1(eb8ad84b772508b6befb35afb11a0d6193c6a060) )
-   ROM_LOAD16_BYTE( "prg0_austria.u22",      0x000000, 0x080000, CRC(46e08afe) SHA1(a6f46581d0f7285704fbf1ac57476c96f4dcbec2) )
-   ROM_LOAD16_BYTE( "prg0_belgium.u22",      0x000000, 0x080000, CRC(f77ab38c) SHA1(8be87175250345d3e31d95ec204805071eae81f6) )
-   ROM_LOAD16_BYTE( "prg0_denmark.u22",      0x000000, 0x080000, CRC(980ca4a2) SHA1(4f29eaa5ba6b94d96c527f80188657abc8f4dcd0) )
-   ROM_LOAD16_BYTE( "prg0_finland.u22",      0x000000, 0x080000, CRC(826d72db) SHA1(be4bca0143f43c13361fd56974eb9b1ce7bd1740) )
-   ROM_LOAD16_BYTE( "prg0_france.u22",       0x000000, 0x080000, CRC(ed1b65f5) SHA1(1e08957c0f7ed65695fb1ceb961ab765f8a97c89) )
-   ROM_LOAD16_BYTE( "prg0_greatbritain.u22", 0x000000, 0x080000, CRC(5c815c87) SHA1(dea89944cd9a3fa6991b214495dc7123a505d39b) )
-   ROM_LOAD16_BYTE( "prg0_greece.u22",       0x000000, 0x080000, CRC(33f74ba9) SHA1(fe770415584b037152b37a75fe468d3c52dcb3cd) )
-   ROM_LOAD16_BYTE( "prg0_netherlands.u22",  0x000000, 0x080000, CRC(e4c42822) SHA1(8bfd286c42d7f2b3c88757b9a8b818be90b73f48) )
-   ROM_LOAD16_BYTE( "prg0_italy.u22",        0x000000, 0x080000, CRC(8bb23f0c) SHA1(b448bba312a8d583a981f6633cbc14af99fdbb06) )
-   ROM_LOAD16_BYTE( "prg0_norway.u22",       0x000000, 0x080000, CRC(3a28067e) SHA1(9435e6ce90b8d740a545469e6edb35d1af11ceab) )
-   ROM_LOAD16_BYTE( "prg0_portugal.u22",     0x000000, 0x080000, CRC(555e1150) SHA1(5c9ae898244a23a4184f9613f42d9aa9530468b9) )
-   ROM_LOAD16_BYTE( "prg0_spain.u22",        0x000000, 0x080000, CRC(0eebaa8c) SHA1(e305e90434e7f322a33e42a642362f770d3eb0e5) )
-   ROM_LOAD16_BYTE( "prg0_sweden.u22",       0x000000, 0x080000, CRC(619dbda2) SHA1(9e88ba104a5cffcced3b93ca711487a82b0fddde) )
-   ROM_LOAD16_BYTE( "prg0_switzerland.u22",  0x000000, 0x080000, CRC(d00784d0) SHA1(0b809414ce910684ca39216086f7d26fd2adeded) )
-   ROM_LOAD16_BYTE( "prg0_australia.u22",    0x000000, 0x080000, CRC(bf7193fe) SHA1(9af50fffc6ef23e300bf7b5e90b0dee6e4f4ad05) )
-   ROM_LOAD16_BYTE( "prg0_newzealand.u22",   0x000000, 0x080000, CRC(6842f075) SHA1(125b303c064d2f0b539ecadcb205756e7fd1334e) )
-   ROM_LOAD16_BYTE( "prg0_taiwan.u22",       0x000000, 0x080000, CRC(0734e75b) SHA1(17a8fb4f8fda3c234ed976490193ba308cac08fe) )
-   ROM_LOAD16_BYTE( "prg0_hongkong.u22",     0x000000, 0x080000, CRC(b6aede29) SHA1(580f29db6a2c2cea43966413778362694992a675) )
-   ROM_LOAD16_BYTE( "prg0_korea.u22",        0x000000, 0x080000, CRC(d9d8c907) SHA1(69c197f2a41f288913f042de9eb8274c0df3ac27) )
-   ROM_LOAD16_BYTE( "prg0_china.u22",        0x000000, 0x080000, CRC(c3b91f7e) SHA1(6b2376c37808dccda296d90ccd7f577ccff4e4dc) )
-   ROM_LOAD16_BYTE( "prg0_none.u22",         0x000000, 0x080000, CRC(accf0850) SHA1(d93e4e80443a40c3a9575dbf21927ef0d1a039b9) )
- */
-
-
-ROM_START( batrider )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "prg0_europe.u22", 0x000000, 0x080000, CRC(91d3e975) SHA1(682885fc17f2424d475c282f239f42faf1aae076) )
-	ROM_LOAD16_BYTE( "prg1b.u23", 0x000001, 0x080000, CRC(8e70b492) SHA1(f84f2039826ae815afb058d71c1dbd190f9d524d) )
-	ROM_LOAD16_BYTE( "prg2.u21" , 0x100000, 0x080000, CRC(bdaa5fbf) SHA1(abd72ac633c0c8e7b4b1d7902c0d6e014ba995fe) )
-	ROM_LOAD16_BYTE( "prg3.u24" , 0x100001, 0x080000, CRC(7aa9f941) SHA1(99bdbad7a96d461073b06a53c50fc57c2fd6fc6d) )
-
-	ROM_REGION( 0x40000, "audiocpu", 0 )            /* Sound Z80 code + bank */
-	ROM_LOAD( "snd.u77", 0x00000, 0x40000, CRC(56682696) SHA1(a372450d9a6d535123dfc31d8116074b168ab646) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "rom-1.bin", 0x000000, 0x400000, CRC(0df69ca2) SHA1(49670347ebd7e1067ff988cf842b275b7ee7b5f7) )
-	ROM_LOAD( "rom-3.bin", 0x400000, 0x400000, CRC(60167d38) SHA1(fd2429808c59ef51fd5f5db84ea89a8dc504186e) )
-	ROM_LOAD( "rom-2.bin", 0x800000, 0x400000, CRC(1bfea593) SHA1(ce06dc3097ae56b0df56d104bbf7efc9b5d968d4) )
-	ROM_LOAD( "rom-4.bin", 0xc00000, 0x400000, CRC(bee03c94) SHA1(5bc1e6769c42857c03456426b502fcb86a114f19) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* ADPCM Samples 1 */
-	ROM_LOAD( "rom-5.bin", 0x000000, 0x100000, CRC(4274daf6) SHA1(85557b4707d529e5914f03c7a856864f5c24950e) )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* ADPCM Samples 2 */
-	ROM_LOAD( "rom-6.bin", 0x000000, 0x100000, CRC(2a1c2426) SHA1(8abc3688ffc5ebb94b8d5118d4fa0908f07fe791) )
-ROM_END
-
-
-ROM_START( batrideru )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "prg0_usa.u22", 0x000000, 0x080000, CRC(2049d007) SHA1(f2a43547a6fc5083b03c1d59a85abbf6e1ce4cd9) )
-	ROM_LOAD16_BYTE( "prg1b.u23", 0x000001, 0x080000, CRC(8e70b492) SHA1(f84f2039826ae815afb058d71c1dbd190f9d524d) )
-	ROM_LOAD16_BYTE( "prg2.u21" , 0x100000, 0x080000, CRC(bdaa5fbf) SHA1(abd72ac633c0c8e7b4b1d7902c0d6e014ba995fe) )
-	ROM_LOAD16_BYTE( "prg3.u24" , 0x100001, 0x080000, CRC(7aa9f941) SHA1(99bdbad7a96d461073b06a53c50fc57c2fd6fc6d) )
-
-	ROM_REGION( 0x40000, "audiocpu", 0 )            /* Sound Z80 code + bank */
-	ROM_LOAD( "snd.u77", 0x00000, 0x40000, CRC(56682696) SHA1(a372450d9a6d535123dfc31d8116074b168ab646) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "rom-1.bin", 0x000000, 0x400000, CRC(0df69ca2) SHA1(49670347ebd7e1067ff988cf842b275b7ee7b5f7) )
-	ROM_LOAD( "rom-3.bin", 0x400000, 0x400000, CRC(60167d38) SHA1(fd2429808c59ef51fd5f5db84ea89a8dc504186e) )
-	ROM_LOAD( "rom-2.bin", 0x800000, 0x400000, CRC(1bfea593) SHA1(ce06dc3097ae56b0df56d104bbf7efc9b5d968d4) )
-	ROM_LOAD( "rom-4.bin", 0xc00000, 0x400000, CRC(bee03c94) SHA1(5bc1e6769c42857c03456426b502fcb86a114f19) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* ADPCM Samples 1 */
-	ROM_LOAD( "rom-5.bin", 0x000000, 0x100000, CRC(4274daf6) SHA1(85557b4707d529e5914f03c7a856864f5c24950e) )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* ADPCM Samples 2 */
-	ROM_LOAD( "rom-6.bin", 0x000000, 0x100000, CRC(2a1c2426) SHA1(8abc3688ffc5ebb94b8d5118d4fa0908f07fe791) )
-ROM_END
-
-
-ROM_START( batriderc )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "prg0_china.u22", 0x000000, 0x080000, CRC(c3b91f7e) SHA1(6b2376c37808dccda296d90ccd7f577ccff4e4dc) )
-	ROM_LOAD16_BYTE( "prg1b.u23", 0x000001, 0x080000, CRC(8e70b492) SHA1(f84f2039826ae815afb058d71c1dbd190f9d524d) )
-	ROM_LOAD16_BYTE( "prg2.u21" , 0x100000, 0x080000, CRC(bdaa5fbf) SHA1(abd72ac633c0c8e7b4b1d7902c0d6e014ba995fe) )
-	ROM_LOAD16_BYTE( "prg3.u24" , 0x100001, 0x080000, CRC(7aa9f941) SHA1(99bdbad7a96d461073b06a53c50fc57c2fd6fc6d) )
-
-	ROM_REGION( 0x40000, "audiocpu", 0 )            /* Sound Z80 code + bank */
-	ROM_LOAD( "snd.u77", 0x00000, 0x40000, CRC(56682696) SHA1(a372450d9a6d535123dfc31d8116074b168ab646) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "rom-1.bin", 0x000000, 0x400000, CRC(0df69ca2) SHA1(49670347ebd7e1067ff988cf842b275b7ee7b5f7) )
-	ROM_LOAD( "rom-3.bin", 0x400000, 0x400000, CRC(60167d38) SHA1(fd2429808c59ef51fd5f5db84ea89a8dc504186e) )
-	ROM_LOAD( "rom-2.bin", 0x800000, 0x400000, CRC(1bfea593) SHA1(ce06dc3097ae56b0df56d104bbf7efc9b5d968d4) )
-	ROM_LOAD( "rom-4.bin", 0xc00000, 0x400000, CRC(bee03c94) SHA1(5bc1e6769c42857c03456426b502fcb86a114f19) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* ADPCM Samples 1 */
-	ROM_LOAD( "rom-5.bin", 0x000000, 0x100000, CRC(4274daf6) SHA1(85557b4707d529e5914f03c7a856864f5c24950e) )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* ADPCM Samples 2 */
-	ROM_LOAD( "rom-6.bin", 0x000000, 0x100000, CRC(2a1c2426) SHA1(8abc3688ffc5ebb94b8d5118d4fa0908f07fe791) )
-ROM_END
-
-
-ROM_START( batriderj )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "prg0b.u22", 0x000000, 0x080000, CRC(4f3fc729) SHA1(b32d51c254741b82171a86c271679522a7aefd34) )
-	ROM_LOAD16_BYTE( "prg1b.u23", 0x000001, 0x080000, CRC(8e70b492) SHA1(f84f2039826ae815afb058d71c1dbd190f9d524d) )
-	ROM_LOAD16_BYTE( "prg2.u21" , 0x100000, 0x080000, CRC(bdaa5fbf) SHA1(abd72ac633c0c8e7b4b1d7902c0d6e014ba995fe) )
-	ROM_LOAD16_BYTE( "prg3.u24" , 0x100001, 0x080000, CRC(7aa9f941) SHA1(99bdbad7a96d461073b06a53c50fc57c2fd6fc6d) )
-
-	ROM_REGION( 0x40000, "audiocpu", 0 )            /* Sound Z80 code + bank */
-	ROM_LOAD( "snd.u77", 0x00000, 0x40000, CRC(56682696) SHA1(a372450d9a6d535123dfc31d8116074b168ab646) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "rom-1.bin", 0x000000, 0x400000, CRC(0df69ca2) SHA1(49670347ebd7e1067ff988cf842b275b7ee7b5f7) )
-	ROM_LOAD( "rom-3.bin", 0x400000, 0x400000, CRC(60167d38) SHA1(fd2429808c59ef51fd5f5db84ea89a8dc504186e) )
-	ROM_LOAD( "rom-2.bin", 0x800000, 0x400000, CRC(1bfea593) SHA1(ce06dc3097ae56b0df56d104bbf7efc9b5d968d4) )
-	ROM_LOAD( "rom-4.bin", 0xc00000, 0x400000, CRC(bee03c94) SHA1(5bc1e6769c42857c03456426b502fcb86a114f19) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* ADPCM Samples 1 */
-	ROM_LOAD( "rom-5.bin", 0x000000, 0x100000, CRC(4274daf6) SHA1(85557b4707d529e5914f03c7a856864f5c24950e) )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* ADPCM Samples 2 */
-	ROM_LOAD( "rom-6.bin", 0x000000, 0x100000, CRC(2a1c2426) SHA1(8abc3688ffc5ebb94b8d5118d4fa0908f07fe791) )
-ROM_END
-
-
-ROM_START( batriderk )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "prg0_korea.u22", 0x000000, 0x080000, CRC(d9d8c907) SHA1(69c197f2a41f288913f042de9eb8274c0df3ac27) )
-	ROM_LOAD16_BYTE( "prg1b.u23", 0x000001, 0x080000, CRC(8e70b492) SHA1(f84f2039826ae815afb058d71c1dbd190f9d524d) )
-	ROM_LOAD16_BYTE( "prg2.u21" , 0x100000, 0x080000, CRC(bdaa5fbf) SHA1(abd72ac633c0c8e7b4b1d7902c0d6e014ba995fe) )
-	ROM_LOAD16_BYTE( "prg3.u24" , 0x100001, 0x080000, CRC(7aa9f941) SHA1(99bdbad7a96d461073b06a53c50fc57c2fd6fc6d) )
-
-	ROM_REGION( 0x40000, "audiocpu", 0 )            /* Sound Z80 code + bank */
-	ROM_LOAD( "snd.u77", 0x00000, 0x40000, CRC(56682696) SHA1(a372450d9a6d535123dfc31d8116074b168ab646) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "rom-1.bin", 0x000000, 0x400000, CRC(0df69ca2) SHA1(49670347ebd7e1067ff988cf842b275b7ee7b5f7) )
-	ROM_LOAD( "rom-3.bin", 0x400000, 0x400000, CRC(60167d38) SHA1(fd2429808c59ef51fd5f5db84ea89a8dc504186e) )
-	ROM_LOAD( "rom-2.bin", 0x800000, 0x400000, CRC(1bfea593) SHA1(ce06dc3097ae56b0df56d104bbf7efc9b5d968d4) )
-	ROM_LOAD( "rom-4.bin", 0xc00000, 0x400000, CRC(bee03c94) SHA1(5bc1e6769c42857c03456426b502fcb86a114f19) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* ADPCM Samples 1 */
-	ROM_LOAD( "rom-5.bin", 0x000000, 0x100000, CRC(4274daf6) SHA1(85557b4707d529e5914f03c7a856864f5c24950e) )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* ADPCM Samples 2 */
-	ROM_LOAD( "rom-6.bin", 0x000000, 0x100000, CRC(2a1c2426) SHA1(8abc3688ffc5ebb94b8d5118d4fa0908f07fe791) )
-ROM_END
-
-/* older version, might have only been released in Japan, Hong Kong and Taiwan? */
-ROM_START( batriderja )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "prg0.bin", 0x000000, 0x080000, CRC(f93ea27c) SHA1(41023c2ee1efd70b5aa9c70e1ddd9e5c3d51d68a) )
-	ROM_LOAD16_BYTE( "prg1.u23", 0x000001, 0x080000, CRC(8ae7f592) SHA1(8a20ebf85eca621f578d2302c3a3988647b077a7) )
-	ROM_LOAD16_BYTE( "prg2.u21", 0x100000, 0x080000, CRC(bdaa5fbf) SHA1(abd72ac633c0c8e7b4b1d7902c0d6e014ba995fe) )
-	ROM_LOAD16_BYTE( "prg3.u24", 0x100001, 0x080000, CRC(7aa9f941) SHA1(99bdbad7a96d461073b06a53c50fc57c2fd6fc6d) )
-
-	ROM_REGION( 0x40000, "audiocpu", 0 )            /* Sound Z80 code + bank */
-	ROM_LOAD( "snd.u77", 0x00000, 0x40000, CRC(56682696) SHA1(a372450d9a6d535123dfc31d8116074b168ab646) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "rom-1.bin", 0x000000, 0x400000, CRC(0df69ca2) SHA1(49670347ebd7e1067ff988cf842b275b7ee7b5f7) )
-	ROM_LOAD( "rom-3.bin", 0x400000, 0x400000, CRC(60167d38) SHA1(fd2429808c59ef51fd5f5db84ea89a8dc504186e) )
-	ROM_LOAD( "rom-2.bin", 0x800000, 0x400000, CRC(1bfea593) SHA1(ce06dc3097ae56b0df56d104bbf7efc9b5d968d4) )
-	ROM_LOAD( "rom-4.bin", 0xc00000, 0x400000, CRC(bee03c94) SHA1(5bc1e6769c42857c03456426b502fcb86a114f19) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* ADPCM Samples 1 */
-	ROM_LOAD( "rom-5.bin", 0x000000, 0x100000, CRC(4274daf6) SHA1(85557b4707d529e5914f03c7a856864f5c24950e) )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* ADPCM Samples 2 */
-	ROM_LOAD( "rom-6.bin", 0x000000, 0x100000, CRC(2a1c2426) SHA1(8abc3688ffc5ebb94b8d5118d4fa0908f07fe791) )
-ROM_END
-
-
-ROM_START( batriderhk )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "prg0.u22", 0x000000, 0x080000, CRC(00afbb7c) SHA1(a4b6331e0fcab7d0c43fc43adb701f1248247b41) )
-	ROM_LOAD16_BYTE( "prg1.u23", 0x000001, 0x080000, CRC(8ae7f592) SHA1(8a20ebf85eca621f578d2302c3a3988647b077a7) )
-	ROM_LOAD16_BYTE( "prg2.u21", 0x100000, 0x080000, CRC(bdaa5fbf) SHA1(abd72ac633c0c8e7b4b1d7902c0d6e014ba995fe) )
-	ROM_LOAD16_BYTE( "prg3.u24", 0x100001, 0x080000, CRC(7aa9f941) SHA1(99bdbad7a96d461073b06a53c50fc57c2fd6fc6d) )
-
-	ROM_REGION( 0x40000, "audiocpu", 0 )            /* Sound Z80 code + bank */
-	ROM_LOAD( "snd.u77", 0x00000, 0x40000, CRC(56682696) SHA1(a372450d9a6d535123dfc31d8116074b168ab646) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "rom-1.bin", 0x000000, 0x400000, CRC(0df69ca2) SHA1(49670347ebd7e1067ff988cf842b275b7ee7b5f7) )
-	ROM_LOAD( "rom-3.bin", 0x400000, 0x400000, CRC(60167d38) SHA1(fd2429808c59ef51fd5f5db84ea89a8dc504186e) )
-	ROM_LOAD( "rom-2.bin", 0x800000, 0x400000, CRC(1bfea593) SHA1(ce06dc3097ae56b0df56d104bbf7efc9b5d968d4) )
-	ROM_LOAD( "rom-4.bin", 0xc00000, 0x400000, CRC(bee03c94) SHA1(5bc1e6769c42857c03456426b502fcb86a114f19) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* ADPCM Samples 1 */
-	ROM_LOAD( "rom-5.bin", 0x000000, 0x100000, CRC(4274daf6) SHA1(85557b4707d529e5914f03c7a856864f5c24950e) )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* ADPCM Samples 2 */
-	ROM_LOAD( "rom-6.bin", 0x000000, 0x100000, CRC(2a1c2426) SHA1(8abc3688ffc5ebb94b8d5118d4fa0908f07fe791) )
-ROM_END
-
-
-ROM_START( batridert )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "u22.bin",  0x000000, 0x080000, CRC(b135820e) SHA1(c222887d18a0a3ea0fcc973b95b29d69c86f7ec3) )
-	ROM_LOAD16_BYTE( "prg1.u23", 0x000001, 0x080000, CRC(8ae7f592) SHA1(8a20ebf85eca621f578d2302c3a3988647b077a7) )
-	ROM_LOAD16_BYTE( "prg2.u21", 0x100000, 0x080000, CRC(bdaa5fbf) SHA1(abd72ac633c0c8e7b4b1d7902c0d6e014ba995fe) )
-	ROM_LOAD16_BYTE( "prg3.u24", 0x100001, 0x080000, CRC(7aa9f941) SHA1(99bdbad7a96d461073b06a53c50fc57c2fd6fc6d) )
-
-	ROM_REGION( 0x40000, "audiocpu", 0 )            /* Sound Z80 code + bank */
-	ROM_LOAD( "snd.u77", 0x00000, 0x40000, CRC(56682696) SHA1(a372450d9a6d535123dfc31d8116074b168ab646) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "rom-1.bin", 0x000000, 0x400000, CRC(0df69ca2) SHA1(49670347ebd7e1067ff988cf842b275b7ee7b5f7) )
-	ROM_LOAD( "rom-3.bin", 0x400000, 0x400000, CRC(60167d38) SHA1(fd2429808c59ef51fd5f5db84ea89a8dc504186e) )
-	ROM_LOAD( "rom-2.bin", 0x800000, 0x400000, CRC(1bfea593) SHA1(ce06dc3097ae56b0df56d104bbf7efc9b5d968d4) )
-	ROM_LOAD( "rom-4.bin", 0xc00000, 0x400000, CRC(bee03c94) SHA1(5bc1e6769c42857c03456426b502fcb86a114f19) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* ADPCM Samples 1 */
-	ROM_LOAD( "rom-5.bin", 0x000000, 0x100000, CRC(4274daf6) SHA1(85557b4707d529e5914f03c7a856864f5c24950e) )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* ADPCM Samples 2 */
-	ROM_LOAD( "rom-6.bin", 0x000000, 0x100000, CRC(2a1c2426) SHA1(8abc3688ffc5ebb94b8d5118d4fa0908f07fe791) )
-ROM_END
-
-
-
-/*
-Battle Bakraid
-Raizing/8ing, 1999
-
-PCB Layout
-----------
-
-ET68-V99
-|-----------------------------------------------------|
-|TA8201  16.93MHz     ROM-6                   6264    |
-|  YAC516                                             |
-|       YMZ280B-F     ROM-7               SND_U0720   |
-|                                                     |
-| VOL                 ROM-8                 Z80       |
-|                                                     |
-|                   341256                            |
-|                                               93C66 |
-|                   341256               XILINX       |
-|J                                       XC95108      |
-|A                  27MHz     32MHz                   |
-|M                                                    |
-|M          DIPSW1                      341256  341256|
-|A                  XILINX    XILINK                  |
-|           DIPSW2  XC95144   XC95108   341256  341256|
-|                                                     |
-|           DIPSW3                                    |
-|                                MACH211    PRG1_U023 |
-| TEST_SW            68000                            |
-|                                           PRG0_U022 |
-|                                                     |
-|                                           PRG3_U024 |
-|                             L7A0498                 |
-|                             GP9001        PRG2_U021 |
-| ROM-0       ROM-1           (QFP208)                |
-|                                                     |
-|                               6264       MN414260   |
-| ROM-2       ROM-3                                   |
-|                               6264       MN414260   |
-|-----------------------------------------------------|
-Notes:
-      ROM-0 to ROM-3 - 32M DIP42
-      ROM-6 to ROM-8 - 32M DIP42 Byte Mode
-      68000 clock - 16.000MHz (32/2)
-      Z80 clock - 5.33333MHz (32/6)
-      VSync - 60Hz
-      HSync - 15.39kHz
-*/
-
-
-
-ROM_START( bbakraid )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "prg0u022_usa.bin", 0x000000, 0x080000, CRC(95fb2ffd) SHA1(c7f502f3945249573b66226e8bacc6a9bc230693) )
-	ROM_LOAD16_BYTE( "prg1u023.new", 0x000001, 0x080000, CRC(4ae9aa64) SHA1(45fdf72141c4c9f24a38d4218c65874799b9c868) )
-	ROM_LOAD16_BYTE( "prg2u021.bin", 0x100000, 0x080000, CRC(ffba8656) SHA1(6526bb65fad3384de3f301a7d1095cbf03757433) )
-	ROM_LOAD16_BYTE( "prg3u024.bin", 0x100001, 0x080000, CRC(834b8ad6) SHA1(0dd6223bb0749819ad29811eeb04fd08d937abb0) )
-
-	ROM_REGION( 0x20000, "audiocpu", 0 )            /* Sound Z80 code */
-	ROM_LOAD( "sndu0720.bin", 0x00000, 0x20000, CRC(e62ab246) SHA1(00d23689dd423ecd4024c58b5903d16e890f1dff) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "gfxu0510.bin", 0x000000, 0x400000, CRC(9cca3446) SHA1(1123f8b8bfbe59a2c572cdf61f1ad27ff37f0f0d) )
-	ROM_LOAD( "gfxu0512.bin", 0x400000, 0x400000, CRC(a2a281d5) SHA1(d9a6623f9433ad682223f9780c26cd1523ebc5c5) )
-	ROM_LOAD( "gfxu0511.bin", 0x800000, 0x400000, CRC(e16472c0) SHA1(6068d679a8b3b65e05acd58a7ce9ead90177049f) )
-	ROM_LOAD( "gfxu0513.bin", 0xc00000, 0x400000, CRC(8bb635a0) SHA1(9064f1a2d8bb88ddbca702fb8556d0dfe6a5cadc) )
-
-	ROM_REGION( 0x0c00000, "ymz", 0 )       /* YMZ280B Samples */
-	ROM_LOAD( "rom6.829", 0x000000, 0x400000, CRC(8848b4a0) SHA1(e0dce136c5d5a4c1a92b863e57848cd5927d06f1) )
-	ROM_LOAD( "rom7.830", 0x400000, 0x400000, CRC(d6224267) SHA1(5c9b7b13effbef9f707811f84bfe50ca85e605e3) )
-	ROM_LOAD( "rom8.831", 0x800000, 0x400000, CRC(a101dfb0) SHA1(4b729b0d562e09df35438e9e6b457b8de2690a6e) )
-
-	ROM_REGION( 0x200, "eeprom", 0 )
-	ROM_LOAD( "eeprom-bbakraid-new.bin", 0x000, 0x200, CRC(35c9275a) SHA1(1282034adf3c7a24545fd273729867058dc93027) )
-ROM_END
-
-
-ROM_START( bbakraidc )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "prg0u022_china.bin", 0x000000, 0x080000, CRC(760be084) SHA1(096c8a2336492d7370ae25f3385faebf6e9c3eca) )
-	ROM_LOAD16_BYTE( "prg1u023.new", 0x000001, 0x080000, CRC(4ae9aa64) SHA1(45fdf72141c4c9f24a38d4218c65874799b9c868) )
-	ROM_LOAD16_BYTE( "prg2u021.bin", 0x100000, 0x080000, CRC(ffba8656) SHA1(6526bb65fad3384de3f301a7d1095cbf03757433) )
-	ROM_LOAD16_BYTE( "prg3u024.bin", 0x100001, 0x080000, CRC(834b8ad6) SHA1(0dd6223bb0749819ad29811eeb04fd08d937abb0) )
-
-	ROM_REGION( 0x20000, "audiocpu", 0 )            /* Sound Z80 code */
-	ROM_LOAD( "sndu0720.bin", 0x00000, 0x20000, CRC(e62ab246) SHA1(00d23689dd423ecd4024c58b5903d16e890f1dff) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "gfxu0510.bin", 0x000000, 0x400000, CRC(9cca3446) SHA1(1123f8b8bfbe59a2c572cdf61f1ad27ff37f0f0d) )
-	ROM_LOAD( "gfxu0512.bin", 0x400000, 0x400000, CRC(a2a281d5) SHA1(d9a6623f9433ad682223f9780c26cd1523ebc5c5) )
-	ROM_LOAD( "gfxu0511.bin", 0x800000, 0x400000, CRC(e16472c0) SHA1(6068d679a8b3b65e05acd58a7ce9ead90177049f) )
-	ROM_LOAD( "gfxu0513.bin", 0xc00000, 0x400000, CRC(8bb635a0) SHA1(9064f1a2d8bb88ddbca702fb8556d0dfe6a5cadc) )
-
-	ROM_REGION( 0x0c00000, "ymz", 0 )       /* YMZ280B Samples */
-	ROM_LOAD( "rom6.829", 0x000000, 0x400000, CRC(8848b4a0) SHA1(e0dce136c5d5a4c1a92b863e57848cd5927d06f1) )
-	ROM_LOAD( "rom7.830", 0x400000, 0x400000, CRC(d6224267) SHA1(5c9b7b13effbef9f707811f84bfe50ca85e605e3) )
-	ROM_LOAD( "rom8.831", 0x800000, 0x400000, CRC(a101dfb0) SHA1(4b729b0d562e09df35438e9e6b457b8de2690a6e) )
-
-	ROM_REGION( 0x200, "eeprom", 0 )
-	ROM_LOAD( "eeprom-bbakraid-new.bin", 0x000, 0x200, CRC(35c9275a) SHA1(1282034adf3c7a24545fd273729867058dc93027) )
-ROM_END
-
-
-ROM_START( bbakraidj )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "prg0u022.new", 0x000000, 0x080000, CRC(fa8d38d3) SHA1(aba91d87a8a62d3fe1139b4437b16e2f844264ad) )
-	ROM_LOAD16_BYTE( "prg1u023.new", 0x000001, 0x080000, CRC(4ae9aa64) SHA1(45fdf72141c4c9f24a38d4218c65874799b9c868) )
-	ROM_LOAD16_BYTE( "prg2u021.bin", 0x100000, 0x080000, CRC(ffba8656) SHA1(6526bb65fad3384de3f301a7d1095cbf03757433) )
-	ROM_LOAD16_BYTE( "prg3u024.bin", 0x100001, 0x080000, CRC(834b8ad6) SHA1(0dd6223bb0749819ad29811eeb04fd08d937abb0) )
-
-	ROM_REGION( 0x20000, "audiocpu", 0 )            /* Sound Z80 code */
-	ROM_LOAD( "sndu0720.bin", 0x00000, 0x20000, CRC(e62ab246) SHA1(00d23689dd423ecd4024c58b5903d16e890f1dff) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "gfxu0510.bin", 0x000000, 0x400000, CRC(9cca3446) SHA1(1123f8b8bfbe59a2c572cdf61f1ad27ff37f0f0d) )
-	ROM_LOAD( "gfxu0512.bin", 0x400000, 0x400000, CRC(a2a281d5) SHA1(d9a6623f9433ad682223f9780c26cd1523ebc5c5) )
-	ROM_LOAD( "gfxu0511.bin", 0x800000, 0x400000, CRC(e16472c0) SHA1(6068d679a8b3b65e05acd58a7ce9ead90177049f) )
-	ROM_LOAD( "gfxu0513.bin", 0xc00000, 0x400000, CRC(8bb635a0) SHA1(9064f1a2d8bb88ddbca702fb8556d0dfe6a5cadc) )
-
-	ROM_REGION( 0x0c00000, "ymz", 0 )       /* YMZ280B Samples */
-	ROM_LOAD( "rom6.829", 0x000000, 0x400000, CRC(8848b4a0) SHA1(e0dce136c5d5a4c1a92b863e57848cd5927d06f1) )
-	ROM_LOAD( "rom7.830", 0x400000, 0x400000, CRC(d6224267) SHA1(5c9b7b13effbef9f707811f84bfe50ca85e605e3) )
-	ROM_LOAD( "rom8.831", 0x800000, 0x400000, CRC(a101dfb0) SHA1(4b729b0d562e09df35438e9e6b457b8de2690a6e) )
-
-	ROM_REGION( 0x200, "eeprom", 0 )
-	ROM_LOAD( "eeprom-bbakraid-new.bin", 0x000, 0x200, CRC(35c9275a) SHA1(1282034adf3c7a24545fd273729867058dc93027) )
-ROM_END
-
-
-ROM_START( bbakraidja )
-	ROM_REGION( 0x200000, "maincpu", 0 )            /* Main 68k code */
-	ROM_LOAD16_BYTE( "prg0u022.bin", 0x000000, 0x080000, CRC(0dd59512) SHA1(c6a4e6aa49c6ac3b04ae62a0a4cc8084ae048381) )
-	ROM_LOAD16_BYTE( "prg1u023.bin", 0x000001, 0x080000, CRC(fecde223) SHA1(eb5ac0eda49b4b0f3d25d8a8bb356e77a453d3a7) )
-	ROM_LOAD16_BYTE( "prg2u021.bin", 0x100000, 0x080000, CRC(ffba8656) SHA1(6526bb65fad3384de3f301a7d1095cbf03757433) )
-	ROM_LOAD16_BYTE( "prg3u024.bin", 0x100001, 0x080000, CRC(834b8ad6) SHA1(0dd6223bb0749819ad29811eeb04fd08d937abb0) )
-
-	ROM_REGION( 0x20000, "audiocpu", 0 )            /* Sound Z80 code */
-	ROM_LOAD( "sndu0720.bin", 0x00000, 0x20000, CRC(e62ab246) SHA1(00d23689dd423ecd4024c58b5903d16e890f1dff) )
-
-	ROM_REGION( 0x1000000, "gp9001", 0 )
-	ROM_LOAD( "gfxu0510.bin", 0x000000, 0x400000, CRC(9cca3446) SHA1(1123f8b8bfbe59a2c572cdf61f1ad27ff37f0f0d) )
-	ROM_LOAD( "gfxu0512.bin", 0x400000, 0x400000, CRC(a2a281d5) SHA1(d9a6623f9433ad682223f9780c26cd1523ebc5c5) )
-	ROM_LOAD( "gfxu0511.bin", 0x800000, 0x400000, CRC(e16472c0) SHA1(6068d679a8b3b65e05acd58a7ce9ead90177049f) )
-	ROM_LOAD( "gfxu0513.bin", 0xc00000, 0x400000, CRC(8bb635a0) SHA1(9064f1a2d8bb88ddbca702fb8556d0dfe6a5cadc) )
-
-	ROM_REGION( 0x0c00000, "ymz", 0 )       /* YMZ280B Samples */
-	ROM_LOAD( "rom6.829", 0x000000, 0x400000, CRC(8848b4a0) SHA1(e0dce136c5d5a4c1a92b863e57848cd5927d06f1) )
-	ROM_LOAD( "rom7.830", 0x400000, 0x400000, CRC(d6224267) SHA1(5c9b7b13effbef9f707811f84bfe50ca85e605e3) )
-	ROM_LOAD( "rom8.831", 0x800000, 0x400000, CRC(a101dfb0) SHA1(4b729b0d562e09df35438e9e6b457b8de2690a6e) )
-
-	ROM_REGION( 0x200, "eeprom", 0 )
-	ROM_LOAD( "eeprom-bbakraid.bin", 0x000, 0x200, CRC(7f97d347) SHA1(3096c399019924dbb7d6673483f6a011f89467c6) )
-ROM_END
-
-
-// dedicated PCB marked Pro Bowl
-ROM_START( nprobowl )
-	ROM_REGION( 0x200000, "maincpu", 0 )
-	ROM_LOAD16_BYTE( "newprobowl-prg0-u021", 0x000000, 0x080000, CRC(3a57f122) SHA1(cc361c295f23bc0479ba49eb15de2ec6ca535a56) ) // 11xxxxxxxxxxxxxxxxx = 0xFF
-	ROM_LOAD16_BYTE( "newprobowl-prg1-u024", 0x000001, 0x080000, CRC(9e9bb58a) SHA1(3d2159bde418dee5d89e3df9a248b4b1989e6ee9) ) // 11xxxxxxxxxxxxxxxxx = 0xFF
-
-	ROM_REGION( 0x200000, "gp9001", 0 )
-	ROM_LOAD16_BYTE( "newprobowl-chr1-u0519", 0x000000, 0x80000, CRC(6700a9bf) SHA1(12a72aa0b91119fbbed994aec702a6869af6f287) )
-	ROM_LOAD16_BYTE( "newprobowl-chr0-u0518", 0x000001, 0x80000, CRC(0736cccd) SHA1(5a4b691be1df697fef3847456c0f4bb3466c403f) )
-	ROM_LOAD16_BYTE( "newprobowl-chr2-u0520", 0x100000, 0x80000, CRC(e5f6d0b6) SHA1(6e1a4792698be4b478118e8c82edb0cf3e2286f2) )
-	ROM_LOAD16_BYTE( "newprobowl-chr3-u0521", 0x100001, 0x80000, CRC(00c21951) SHA1(922abde172fb82b504dce41b95227740f16208a7) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )
-	ROM_LOAD( "newprobowl-adpcm0-u0834", 0x00000, 0x80000, CRC(3b40b161) SHA1(ff8ba38dd7e0dadbf72810470e3d9afb1cd983d2) )
-	ROM_LOAD( "newprobowl-adpcm1-u0835", 0x80000, 0x80000, CRC(8c191e60) SHA1(f81c2849ffc553d921fc680cd50c2997b834c44a) )
-ROM_END
-
-ROM_START( probowl2 ) // identical to New Pro Bowl but for slight mods to the GFX ROMs to display the different title
-	ROM_REGION( 0x200000, "maincpu", 0 )
-	ROM_LOAD16_BYTE( "probowl2.prg0.u021", 0x000000, 0x080000, CRC(3a57f122) SHA1(cc361c295f23bc0479ba49eb15de2ec6ca535a56) ) // 11xxxxxxxxxxxxxxxxx = 0xFF
-	ROM_LOAD16_BYTE( "probowl2.prg1.u024", 0x000001, 0x080000, CRC(9e9bb58a) SHA1(3d2159bde418dee5d89e3df9a248b4b1989e6ee9) ) // 11xxxxxxxxxxxxxxxxx = 0xFF
-
-	ROM_REGION( 0x200000, "gp9001", 0 )
-	ROM_LOAD16_BYTE( "probowl2.chr1.u0519", 0x000000, 0x80000, CRC(66a5c9cd) SHA1(b8edc6a66e3929af2a1399e5fb6fc341b9789136) )
-	ROM_LOAD16_BYTE( "probowl2.chr0.u0518", 0x000001, 0x80000, CRC(b4439673) SHA1(5ca19d9c3c3cce9485fb9bd5224f0b14abe07f49) )
-	ROM_LOAD16_BYTE( "probowl2.chr2.u0520", 0x100000, 0x80000, CRC(e05be3a0) SHA1(922c6f094358cf3df790063e68d5dd4a32d46b06) )
-	ROM_LOAD16_BYTE( "probowl2.chr3.u0521", 0x100001, 0x80000, CRC(55e2565f) SHA1(b6a570b736a9b30e26d2a59b53f218ef5cd6f0f6) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )
-	ROM_LOAD( "probowl2.adpcm0.u0834", 0x00000, 0x80000, CRC(be2fb43f) SHA1(f3f9bc522f4668852b751f291cef0000bb86b779) )
-	ROM_LOAD( "probowl2.adpcm1.u0835", 0x80000, 0x80000, CRC(8c191e60) SHA1(f81c2849ffc553d921fc680cd50c2997b834c44a) )
-ROM_END
-
-
 
 
 GAME( 1993, sstriker,    0,        mahoudai,   sstriker,   sstriker_state, empty_init,      ROT270, "Raizing",                         "Sorcer Striker",           MACHINE_SUPPORTS_SAVE ) // verified on two different PCBs
@@ -2962,26 +1579,3 @@ GAME( 1996, bgareggacn,  bgaregga, bgaregga,   bgareggacn, bgaregga_state, init_
 GAME( 1998, bgareggabl,  bgaregga, bgareggabl, bgareggabl, bgaregga_bootleg_state, init_bgaregga,   ROT270, "bootleg (Melody)",   "1945 Er Dai / 1945 Part-2 (Chinese hack of Battle Garegga)",               MACHINE_SUPPORTS_SAVE ) // based on Thu Feb 1 1996 set, Region hardcoded to China
 GAME( 1997, bgareggabla, bgaregga, bgareggabl, bgareggabl, bgaregga_bootleg_state, init_bgaregga,   ROT270, "bootleg (Melody)",   "Leishen Chuan / Thunder Deity Biography (Chinese hack of Battle Garegga)", MACHINE_SUPPORTS_SAVE ) // based on Thu Feb 1 1996 set, Region hardcoded to Asia
 GAME( 1996, bgareggablj, bgaregga, bgareggabl, bgareggabl, bgaregga_bootleg_state, init_bgaregga,   ROT270, "bootleg",            "Battle Garegga (Japan, bootleg) (Sat Feb 3 1996)",                         MACHINE_SUPPORTS_SAVE )
-
-// these are all based on Version B, even if only the Japan version states 'version B'
-GAME( 1998, batrider,    0,        batrider,   batrider,   batrider_state, init_batrider,   ROT270, "Raizing / Eighting", "Armed Police Batrider (Europe) (Fri Feb 13 1998)",           MACHINE_SUPPORTS_SAVE )
-GAME( 1998, batrideru,   batrider, batrider,   batrider,   batrider_state, init_batrider,   ROT270, "Raizing / Eighting", "Armed Police Batrider (USA) (Fri Feb 13 1998)",              MACHINE_SUPPORTS_SAVE )
-GAME( 1998, batriderc,   batrider, batrider,   batrider,   batrider_state, init_batrider,   ROT270, "Raizing / Eighting", "Armed Police Batrider (China) (Fri Feb 13 1998)",            MACHINE_SUPPORTS_SAVE )
-GAME( 1998, batriderj,   batrider, batrider,   batriderj,  batrider_state, init_batrider,   ROT270, "Raizing / Eighting", "Armed Police Batrider (Japan, B version) (Fri Feb 13 1998)", MACHINE_SUPPORTS_SAVE )
-GAME( 1998, batriderk,   batrider, batrider,   batrider,   batrider_state, init_batrider,   ROT270, "Raizing / Eighting", "Armed Police Batrider (Korea) (Fri Feb 13 1998)",            MACHINE_SUPPORTS_SAVE )
-// older revision of the code
-GAME( 1998, batriderja,  batrider, batrider,   batriderj,  batrider_state, init_batrider,   ROT270, "Raizing / Eighting", "Armed Police Batrider (Japan, older version) (Mon Dec 22 1997)", MACHINE_SUPPORTS_SAVE )
-GAME( 1998, batriderhk,  batrider, batrider,   batrider,   batrider_state, init_batrider,   ROT270, "Raizing / Eighting", "Armed Police Batrider (Hong Kong) (Mon Dec 22 1997)",            MACHINE_SUPPORTS_SAVE )
-GAME( 1998, batridert,   batrider, batrider,   batrider,   batrider_state, init_batrider,   ROT270, "Raizing / Eighting", "Armed Police Batrider (Taiwan) (Mon Dec 22 1997)",               MACHINE_SUPPORTS_SAVE )
-
-// Battle Bakraid
-// the 'unlimited' version is a newer revision of the code
-GAME( 1999, bbakraid,    0,        bbakraid,   bbakraid,   bbakraid_state, init_bbakraid,   ROT270, "Eighting", "Battle Bakraid - Unlimited Version (USA) (Tue Jun 8 1999)",   MACHINE_SUPPORTS_SAVE )
-GAME( 1999, bbakraidc,   bbakraid, bbakraid,   bbakraid,   bbakraid_state, init_bbakraid,   ROT270, "Eighting", "Battle Bakraid - Unlimited Version (China) (Tue Jun 8 1999)", MACHINE_SUPPORTS_SAVE )
-GAME( 1999, bbakraidj,   bbakraid, bbakraid,   bbakraid,   bbakraid_state, init_bbakraid,   ROT270, "Eighting", "Battle Bakraid - Unlimited Version (Japan) (Tue Jun 8 1999)", MACHINE_SUPPORTS_SAVE )
-// older revision of the code
-GAME( 1999, bbakraidja,  bbakraid, bbakraid,   bbakraid,   bbakraid_state, init_bbakraid,   ROT270, "Eighting", "Battle Bakraid (Japan) (Wed Apr 7 1999)", MACHINE_SUPPORTS_SAVE )
-
-// dedicated PCB
-GAME( 1996, nprobowl,    0,        nprobowl,   nprobowl,   nprobowl_state, empty_init,      ROT0,   "Zuck / Able Corp", "New Pro Bowl", MACHINE_NOT_WORKING | MACHINE_IMPERFECT_SOUND | MACHINE_MECHANICAL | MACHINE_SUPPORTS_SAVE ) // bad GFXs, no sound banking, controls, etc
-GAME( 1996, probowl2,    nprobowl, nprobowl,   nprobowl,   nprobowl_state, empty_init,      ROT0,   "Zuck / Able Corp", "Pro Bowl 2",   MACHINE_NOT_WORKING | MACHINE_IMPERFECT_SOUND | MACHINE_MECHANICAL | MACHINE_SUPPORTS_SAVE ) // bad GFXs, no sound banking, controls, etc
